Move transform worker support classes into their own source files
Keep each top-level type in its own file so later pipeline extraction can
land under the 500 SLOC limit without a partial-class split. Type bodies
are unchanged; NameofRules stays nested in TransformWorkerProgram.
