+++ Packages/src/Editor/FirstPartyTools/HotReload/TransformWorker~/PatchKinds.cs
@@ -1,0 +1,25 @@
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Collections.Immutable;
+using System.Globalization;
+using System.IO;
+using System.Linq;
+using System.Runtime.Loader;
+using System.Security.Cryptography;
+using System.Text;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Text;
+
+internal static class PatchKinds
+{
+    public const string Transplant = "transplant";
+    public const string Delegation = "delegation";
+
+    // Keep in sync with HotReloadConstants.PatchKindAddedMethod.
+    public const string AddedMethod = "addedMethod";
+}
