+++ Packages/src/Editor/FirstPartyTools/HotReload/TransformWorker~/RemovedMemberKinds.cs
@@ -1,0 +1,24 @@
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Collections.Immutable;
+using System.Globalization;
+using System.IO;
+using System.Linq;
+using System.Runtime.Loader;
+using System.Security.Cryptography;
+using System.Text;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Text;
+
+internal static class RemovedMemberKinds
+{
+    // Keep in sync with HotReloadConstants.RemovedMemberKindMethod / RemovedMemberKindField.
+    public const string Method = "method";
+
+    public const string Field = "field";
+}
