+++ Packages/src/Editor/FirstPartyTools/HotReload/TransformWorker~/TransformWorker.cs
@@ -5955,3480 +5955,3 @@ public static class TransformWorkerProgram
         return leftName == rightName;
     }
 }
-
-internal static class AccessibilityRules
-{
-    public static bool IsInaccessibleFromExternalAssembly(ISymbol symbol)
-    {
-        if (IsNonCrossAssemblySymbol(symbol))
-        {
-            return false;
-        }
-
-        // Local functions / lambdas are emitted into the shim assembly itself, so they have no
-        // cross-assembly accessibility problem and must not be treated as accessor targets.
-        if (IsLocalOrAnonymousFunction(symbol))
-        {
-            return false;
-        }
-
-        if (symbol is ITypeSymbol typeSymbol)
-        {
-            return HasInaccessibleAccessibility(typeSymbol.DeclaredAccessibility)
-                || (typeSymbol.ContainingType != null
-                    && IsInaccessibleFromExternalAssembly(typeSymbol.ContainingType));
-        }
-
-        if (IsMemberSymbol(symbol))
-        {
-            return HasInaccessibleMemberAccessibility(symbol);
-        }
-
-        return false;
-    }
-
-    private static bool IsNonCrossAssemblySymbol(ISymbol symbol)
-    {
-        return symbol is ILocalSymbol
-            || symbol is IParameterSymbol
-            || symbol is IRangeVariableSymbol
-            || symbol is ITypeParameterSymbol
-            || symbol is INamespaceSymbol
-            || symbol is ILabelSymbol
-            || symbol is IDiscardSymbol;
-    }
-
-    private static bool IsLocalOrAnonymousFunction(ISymbol symbol)
-    {
-        return symbol is IMethodSymbol methodKindSymbol
-            && (methodKindSymbol.MethodKind == MethodKind.LocalFunction
-                || methodKindSymbol.MethodKind == MethodKind.AnonymousFunction);
-    }
-
-    private static bool IsMemberSymbol(ISymbol symbol)
-    {
-        return symbol is IFieldSymbol
-            || symbol is IPropertySymbol
-            || symbol is IMethodSymbol
-            || symbol is IEventSymbol;
-    }
-
-    private static bool HasInaccessibleMemberAccessibility(ISymbol symbol)
-    {
-        if (HasInaccessibleAccessibility(symbol.DeclaredAccessibility))
-        {
-            return true;
-        }
-
-        // Recurse through nested containing types (same rule as the type-symbol branch).
-        return symbol.ContainingType != null
-            && IsInaccessibleFromExternalAssembly(symbol.ContainingType);
-    }
-
-    /// <summary>
-    /// What: accessibility of a property get/set accessor method (not the property declaration).
-    /// Missing accessors are treated as inaccessible so write-only/read-only misuse fails closed.
-    /// </summary>
-    public static bool IsInaccessibleAccessor(IMethodSymbol accessorMethod)
-    {
-        if (accessorMethod == null)
-        {
-            return true;
-        }
-
-        return IsInaccessibleFromExternalAssembly(accessorMethod);
-    }
-
-    private static bool HasInaccessibleAccessibility(Accessibility accessibility)
-    {
-        return accessibility == Accessibility.Private
-            || accessibility == Accessibility.Internal
-            || accessibility == Accessibility.Protected
-            || accessibility == Accessibility.ProtectedAndInternal
-            || accessibility == Accessibility.ProtectedOrInternal;
-    }
-
-    /// <summary>
-    /// What: whether a type can appear in an accessor signature / as a type mention in a shim
-    /// that will JIT outside Harmony skip-visibility.
-    /// </summary>
-    public static bool IsExternallyVisibleType(ITypeSymbol typeSymbol)
-    {
-        if (typeSymbol == null || typeSymbol.TypeKind == TypeKind.Error)
-        {
-            return false;
-        }
-
-        if (typeSymbol is ITypeParameterSymbol)
-        {
-            return true;
-        }
-
-        if (typeSymbol is IArrayTypeSymbol arrayType)
-        {
-            return IsExternallyVisibleType(arrayType.ElementType);
-        }
-
-        if (typeSymbol is IPointerTypeSymbol pointerType)
-        {
-            return IsExternallyVisibleType(pointerType.PointedAtType);
-        }
-
-        if (typeSymbol is INamedTypeSymbol namedType)
-        {
-            if (IsInaccessibleFromExternalAssembly(namedType))
-            {
-                return false;
-            }
-
-            foreach (ITypeSymbol typeArgument in namedType.TypeArguments)
-            {
-                if (!IsExternallyVisibleType(typeArgument))
-                {
-                    return false;
-                }
-            }
-
-            return true;
-        }
-
-        return !IsInaccessibleFromExternalAssembly(typeSymbol);
-    }
-}
-
-/// <summary>
-/// What: per-shim-type registry of Harmony accessor delegates to emit and bind.
-/// </summary>
-internal sealed class AccessorPlan
-{
-    private readonly List<AccessorEntry> _entries = new List<AccessorEntry>();
-    private readonly Dictionary<string, AccessorEntry> _byKey =
-        new Dictionary<string, AccessorEntry>(StringComparer.Ordinal);
-
-    public IReadOnlyList<AccessorEntry> Entries => _entries;
-
-    public AccessorEntry GetOrAddField(IFieldSymbol fieldSymbol)
-    {
-        string key = "F:" + BuildMemberKey(fieldSymbol);
-        if (_byKey.TryGetValue(key, out AccessorEntry existing))
-        {
-            return existing;
-        }
-
-        string fieldName = AllocateName("__F_" + SanitizeIdentifier(fieldSymbol.Name));
-        AccessorEntry entry = AccessorEntry.ForField(fieldSymbol, fieldName, key);
-        _byKey[key] = entry;
-        _entries.Add(entry);
-        return entry;
-    }
-
-    public AccessorEntry GetOrAddMethod(IMethodSymbol methodSymbol)
-    {
-        string key = "M:" + BuildMemberKey(methodSymbol);
-        if (_byKey.TryGetValue(key, out AccessorEntry existing))
-        {
-            return existing;
-        }
-
-        string fieldName = AllocateName("__M_" + SanitizeIdentifier(methodSymbol.Name));
-        AccessorEntry entry = AccessorEntry.ForMethod(methodSymbol, fieldName, key);
-        _byKey[key] = entry;
-        _entries.Add(entry);
-        return entry;
-    }
-
-    public AccessorEntry GetOrAddPropertyGetter(IPropertySymbol propertySymbol)
-    {
-        string key = "PG:" + BuildMemberKey(propertySymbol);
-        if (_byKey.TryGetValue(key, out AccessorEntry existing))
-        {
-            return existing;
-        }
-
-        string fieldName = AllocateName("__P_get_" + SanitizeIdentifier(propertySymbol.Name));
-        AccessorEntry entry = AccessorEntry.ForPropertyGetter(propertySymbol, fieldName, key);
-        _byKey[key] = entry;
-        _entries.Add(entry);
-        return entry;
-    }
-
-    public AccessorEntry GetOrAddPropertySetter(IPropertySymbol propertySymbol)
-    {
-        string key = "PS:" + BuildMemberKey(propertySymbol);
-        if (_byKey.TryGetValue(key, out AccessorEntry existing))
-        {
-            return existing;
-        }
-
-        string fieldName = AllocateName("__P_set_" + SanitizeIdentifier(propertySymbol.Name));
-        AccessorEntry entry = AccessorEntry.ForPropertySetter(propertySymbol, fieldName, key);
-        _byKey[key] = entry;
-        _entries.Add(entry);
-        return entry;
-    }
-
-    private string AllocateName(string preferred)
-    {
-        if (!_entries.Any(entry => entry.DelegateFieldName == preferred))
-        {
-            return preferred;
-        }
-
-        int suffix = 2;
-        while (_entries.Any(entry => entry.DelegateFieldName == preferred + suffix))
-        {
-            suffix++;
-        }
-
-        return preferred + suffix;
-    }
-
-    /// <summary>
-    /// What: stable identity for a member across overloads and same-named members on different
-    /// types — containing type FQ + name + (parameter type FQs).
-    /// </summary>
-    public static string BuildMemberKey(ISymbol symbol)
-    {
-        string typePart = symbol.ContainingType == null
-            ? string.Empty
-            : symbol.ContainingType.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat);
-        string name = symbol.Name;
-
-        if (symbol is IMethodSymbol methodSymbol)
-        {
-            string args = string.Join(
-                ",",
-                methodSymbol.Parameters.Select(
-                    parameter => parameter.Type.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat)));
-            return typePart + "." + name + "(" + args + ")";
-        }
-
-        if (symbol is IPropertySymbol propertySymbol)
-        {
-            string args = string.Join(
-                ",",
-                propertySymbol.Parameters.Select(
-                    parameter => parameter.Type.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat)));
-            return typePart + "." + name + "(" + args + ")";
-        }
-
-        return typePart + "." + name + "()";
-    }
-
-    private static string SanitizeIdentifier(string name)
-    {
-        if (string.IsNullOrEmpty(name))
-        {
-            return "member";
-        }
-
-        StringBuilder builder = new StringBuilder(name.Length);
-        foreach (char character in name)
-        {
-            if (char.IsLetterOrDigit(character) || character == '_')
-            {
-                builder.Append(character);
-            }
-            else
-            {
-                builder.Append('_');
-            }
-        }
-
-        return builder.Length == 0 ? "member" : builder.ToString();
-    }
-}
-
-internal enum AccessorKind
-{
-    FieldRef,
-    MethodDelegate,
-    PropertyGetter,
-    PropertySetter
-}
-
-/// <summary>
-/// What: one Harmony accessor delegate field plus the statements that bind it in __BindAccessors.
-/// </summary>
-internal sealed class AccessorEntry
-{
-    public AccessorKind Kind { get; private set; }
-
-    public string RegistryKey { get; private set; }
-
-    public string DelegateFieldName { get; private set; }
-
-    public IFieldSymbol FieldSymbol { get; private set; }
-
-    public IMethodSymbol MethodSymbol { get; private set; }
-
-    public IPropertySymbol PropertySymbol { get; private set; }
-
-    public static AccessorEntry ForField(
-        IFieldSymbol fieldSymbol,
-        string delegateFieldName,
-        string registryKey)
-    {
-        return new AccessorEntry
-        {
-            Kind = AccessorKind.FieldRef,
-            RegistryKey = registryKey,
-            DelegateFieldName = delegateFieldName,
-            FieldSymbol = fieldSymbol
-        };
-    }
-
-    public static AccessorEntry ForMethod(
-        IMethodSymbol methodSymbol,
-        string delegateFieldName,
-        string registryKey)
-    {
-        return new AccessorEntry
-        {
-            Kind = AccessorKind.MethodDelegate,
-            RegistryKey = registryKey,
-            DelegateFieldName = delegateFieldName,
-            MethodSymbol = methodSymbol
-        };
-    }
-
-    public static AccessorEntry ForPropertyGetter(
-        IPropertySymbol propertySymbol,
-        string delegateFieldName,
-        string registryKey)
-    {
-        return new AccessorEntry
-        {
-            Kind = AccessorKind.PropertyGetter,
-            RegistryKey = registryKey,
-            DelegateFieldName = delegateFieldName,
-            PropertySymbol = propertySymbol
-        };
-    }
-
-    public static AccessorEntry ForPropertySetter(
-        IPropertySymbol propertySymbol,
-        string delegateFieldName,
-        string registryKey)
-    {
-        return new AccessorEntry
-        {
-            Kind = AccessorKind.PropertySetter,
-            RegistryKey = registryKey,
-            DelegateFieldName = delegateFieldName,
-            PropertySymbol = propertySymbol
-        };
-    }
-
-    public bool TryGetVisibilityFailure(out string reason)
-    {
-        foreach (ITypeSymbol typeSymbol in EnumerateSignatureTypes())
-        {
-            if (!AccessibilityRules.IsExternallyVisibleType(typeSymbol))
-            {
-                reason = "accessor signature type is not visible from an external assembly: "
-                    + typeSymbol.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat);
-                return true;
-            }
-        }
-
-        reason = null;
-        return false;
-    }
-
-    public IEnumerable<ITypeSymbol> EnumerateSignatureTypes()
-    {
-        // Bind statements always emit typeof(ContainingType), including for static members.
-        switch (Kind)
-        {
-            case AccessorKind.FieldRef:
-                yield return FieldSymbol.ContainingType;
-                yield return FieldSymbol.Type;
-                yield break;
-            case AccessorKind.MethodDelegate:
-                yield return MethodSymbol.ContainingType;
-                foreach (IParameterSymbol parameter in MethodSymbol.Parameters)
-                {
-                    yield return parameter.Type;
-                }
-
-                if (!MethodSymbol.ReturnsVoid)
-                {
-                    yield return MethodSymbol.ReturnType;
-                }
-
-                yield break;
-            case AccessorKind.PropertyGetter:
-            case AccessorKind.PropertySetter:
-                yield return PropertySymbol.ContainingType;
-                yield return PropertySymbol.Type;
-                yield break;
-        }
-    }
-
-    public FieldDeclarationSyntax EmitFieldDeclaration()
-    {
-        TypeSyntax fieldType = SyntaxFactory.ParseTypeName(BuildDelegateTypeDisplayString());
-        return SyntaxFactory.FieldDeclaration(
-                SyntaxFactory.VariableDeclaration(fieldType)
-                    .WithVariables(
-                        SyntaxFactory.SingletonSeparatedList(
-                            SyntaxFactory.VariableDeclarator(DelegateFieldName))))
-            .WithModifiers(
-                SyntaxFactory.TokenList(
-                    SyntaxFactory.Token(SyntaxKind.PublicKeyword),
-                    SyntaxFactory.Token(SyntaxKind.StaticKeyword)));
-    }
-
-    public StatementSyntax EmitBindStatement()
-    {
-        string statement = Kind switch
-        {
-            AccessorKind.FieldRef => BuildFieldRefBindStatement(),
-            AccessorKind.MethodDelegate => BuildMethodDelegateBindStatement(
-                MethodSymbol.Name,
-                MethodSymbol),
-            AccessorKind.PropertyGetter => BuildMethodDelegateBindStatement(
-                PropertySymbol.GetMethod.Name,
-                PropertySymbol.GetMethod),
-            AccessorKind.PropertySetter => BuildMethodDelegateBindStatement(
-                PropertySymbol.SetMethod.Name,
-                PropertySymbol.SetMethod),
-            _ => throw new InvalidOperationException("Unknown accessor kind.")
-        };
-
-        return SyntaxFactory.ParseStatement(statement);
-    }
-
-    private string BuildDelegateTypeDisplayString()
-    {
-        switch (Kind)
-        {
-            case AccessorKind.FieldRef:
-                if (FieldSymbol.IsStatic)
-                {
-                    return "global::HarmonyLib.AccessTools.FieldRef<"
-                        + TypeDisplay(FieldSymbol.Type) + ">";
-                }
-
-                return "global::HarmonyLib.AccessTools.FieldRef<"
-                    + TypeDisplay(FieldSymbol.ContainingType) + ", "
-                    + TypeDisplay(FieldSymbol.Type) + ">";
-            case AccessorKind.MethodDelegate:
-                return BuildFuncOrActionType(MethodSymbol);
-            case AccessorKind.PropertyGetter:
-                return BuildFuncOrActionType(PropertySymbol.GetMethod);
-            case AccessorKind.PropertySetter:
-                return BuildFuncOrActionType(PropertySymbol.SetMethod);
-            default:
-                throw new InvalidOperationException("Unknown accessor kind.");
-        }
-    }
-
-    private static string BuildFuncOrActionType(IMethodSymbol methodSymbol)
-    {
-        List<string> typeArguments = new List<string>();
-        foreach (ITypeSymbol parameterType in EnumerateDelegateParameterTypes(methodSymbol))
-        {
-            typeArguments.Add(TypeDisplay(parameterType));
-        }
-
-        if (methodSymbol.ReturnsVoid)
-        {
-            if (typeArguments.Count == 0)
-            {
-                return "global::System.Action";
-            }
-
-            return "global::System.Action<" + string.Join(", ", typeArguments) + ">";
-        }
-
-        typeArguments.Add(TypeDisplay(methodSymbol.ReturnType));
-        return "global::System.Func<" + string.Join(", ", typeArguments) + ">";
-    }
-
-    private string BuildFieldRefBindStatement()
-    {
-        if (FieldSymbol.IsStatic)
-        {
-            // Why FieldInfo: the Type+name StaticFieldRefAccess overloads return ref F,
-            // not a FieldRef`1 that __BindAccessors can store.
-            return DelegateFieldName + " = global::HarmonyLib.AccessTools.StaticFieldRefAccess<"
-                + TypeDisplay(FieldSymbol.Type)
-                + ">(global::HarmonyLib.AccessTools.Field(typeof("
-                + TypeDisplay(FieldSymbol.ContainingType) + "), \""
-                + EscapeStringLiteral(FieldSymbol.Name) + "\"));";
-        }
-
-        return DelegateFieldName + " = global::HarmonyLib.AccessTools.FieldRefAccess<"
-            + TypeDisplay(FieldSymbol.ContainingType) + ", "
-            + TypeDisplay(FieldSymbol.Type) + ">(\""
-            + EscapeStringLiteral(FieldSymbol.Name) + "\");";
-    }
-
-    private string BuildMethodDelegateBindStatement(string metadataName, IMethodSymbol methodSymbol)
-    {
-        string declaringType = TypeDisplay(methodSymbol.ContainingType);
-        string delegateType = BuildFuncOrActionType(methodSymbol);
-        List<ITypeSymbol> delegateParameterTypes = EnumerateDelegateParameterTypes(methodSymbol);
-        // AccessTools.Method matches metadata parameters only; the instance receiver is not one.
-        IReadOnlyList<ITypeSymbol> methodLookupTypes = methodSymbol.IsStatic
-            ? delegateParameterTypes
-            : delegateParameterTypes.GetRange(1, delegateParameterTypes.Count - 1);
-        string typeArray = BuildTypeArrayLiteral(methodLookupTypes);
-        // virtualCall must stay true for virtual/override/abstract instance members so a derived
-        // override is dispatched; non-virtual private/internal targets keep false (exact method).
-        bool virtualCall = !methodSymbol.IsStatic
-            && (methodSymbol.IsVirtual || methodSymbol.IsOverride || methodSymbol.IsAbstract);
-        string virtualCallLiteral = virtualCall ? "true" : "false";
-        // Why not null: Harmony then uses Func<> generic arguments including TResult as
-        // DynamicMethod parameters, so Func<Host,T> becomes T(Host,T) and bind fails.
-        string delegateArgs = BuildTypeArrayLiteral(delegateParameterTypes);
-        return DelegateFieldName + " = global::HarmonyLib.AccessTools.MethodDelegate<"
-            + delegateType + ">(global::HarmonyLib.AccessTools.Method(typeof("
-            + declaringType + "), \"" + EscapeStringLiteral(metadataName) + "\", "
-            + typeArray + "), null, " + virtualCallLiteral + ", " + delegateArgs + ");";
-    }
-
-    // Open-delegate parameter types: declaring type first for instance methods, then each
-    // method parameter. Excludes Func TResult so Harmony arity matches the delegate Invoke.
-    private static List<ITypeSymbol> EnumerateDelegateParameterTypes(IMethodSymbol methodSymbol)
-    {
-        List<ITypeSymbol> types = new List<ITypeSymbol>();
-        if (!methodSymbol.IsStatic)
-        {
-            types.Add(methodSymbol.ContainingType);
-        }
-
-        foreach (IParameterSymbol parameter in methodSymbol.Parameters)
-        {
-            types.Add(parameter.Type);
-        }
-
-        return types;
-    }
-
-    private static string BuildTypeArrayLiteral(IReadOnlyList<ITypeSymbol> types)
-    {
-        if (types.Count == 0)
-        {
-            return "new global::System.Type[] { }";
-        }
-
-        IEnumerable<string> typeofs = types.Select(type => "typeof(" + TypeDisplay(type) + ")");
-        return "new global::System.Type[] { " + string.Join(", ", typeofs) + " }";
-    }
-
-    private static string TypeDisplay(ITypeSymbol typeSymbol)
-    {
-        return typeSymbol.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat);
-    }
-
-    private static string EscapeStringLiteral(string value)
-    {
-        return value.Replace("\\", "\\\\").Replace("\"", "\\\"");
-    }
-}
-
-/// <summary>
-/// What: decides whether an async/iterator/closure private-access skip can be rescued by
-/// rewriting inaccessible member accesses into Harmony accessor delegates (conditions b/c).
-/// </summary>
-internal static class AccessorEligibility
-{
-    public static bool TryBuildPlan(
-        SemanticModel semanticModel,
-        IMethodSymbol methodSymbol,
-        INamedTypeSymbol typeSymbol,
-        SyntaxNode bodyNode,
-        out AccessorPlan plan,
-        out string rejectReason)
-    {
-        plan = null;
-        rejectReason = null;
-
-        if (!AccessibilityRules.IsExternallyVisibleType(typeSymbol))
-        {
-            rejectReason = "containing type is not visible from an external assembly (condition c).";
-            return false;
-        }
-
-        if (!AreMethodSignatureTypesVisible(methodSymbol, out rejectReason))
-        {
-            return false;
-        }
-
-        if (!AreBodyTypeUsagesVisible(semanticModel, bodyNode, out rejectReason))
-        {
-            return false;
-        }
-
-        AccessorPlan built = new AccessorPlan();
-        foreach (SyntaxNode node in bodyNode.DescendantNodesAndSelf())
-        {
-            if (TransformWorkerProgram.NameofRules.IsInsideNameofArgument(node))
-            {
-                continue;
-            }
-
-            if (!TryRegisterInaccessibleAccess(semanticModel, node, built, out rejectReason))
-            {
-                if (rejectReason != null)
-                {
-                    return false;
-                }
-            }
-        }
-
-        foreach (AccessorEntry entry in built.Entries)
-        {
-            if (entry.TryGetVisibilityFailure(out rejectReason))
-            {
-                rejectReason = rejectReason + " (condition c).";
-                return false;
-            }
-        }
-
-        if (NeedsPropertyIncrementRewrite(semanticModel, bodyNode))
-        {
-            rejectReason =
-                "inaccessible property increment/decrement has no accessor rewrite shape.";
-            return false;
-        }
-
-        plan = built;
-        rejectReason = null;
-        return true;
-    }
-
-    private static bool AreMethodSignatureTypesVisible(IMethodSymbol methodSymbol, out string rejectReason)
-    {
-        if (!AccessibilityRules.IsExternallyVisibleType(methodSymbol.ReturnType))
-        {
-            rejectReason = "method return type is not visible from an external assembly (condition c).";
-            return false;
-        }
-
-        foreach (IParameterSymbol parameter in methodSymbol.Parameters)
-        {
-            if (!AccessibilityRules.IsExternallyVisibleType(parameter.Type))
-            {
-                rejectReason =
-                    "method parameter type is not visible from an external assembly (condition c).";
-                return false;
-            }
-        }
-
-        rejectReason = null;
-        return true;
-    }
-
-    private static bool AreBodyTypeUsagesVisible(
-        SemanticModel semanticModel,
-        SyntaxNode bodyNode,
-        out string rejectReason)
-    {
-        foreach (SyntaxNode node in bodyNode.DescendantNodesAndSelf())
-        {
-            if (TransformWorkerProgram.NameofRules.IsInsideNameofArgument(node))
-            {
-                continue;
-            }
-
-            ITypeSymbol typeSymbol = null;
-            if (node is TypeSyntax typeSyntax)
-            {
-                typeSymbol = semanticModel.GetTypeInfo(typeSyntax).Type
-                    ?? semanticModel.GetSymbolInfo(typeSyntax).Symbol as ITypeSymbol;
-            }
-            else if (node is VariableDeclarationSyntax variableDeclaration
-                && variableDeclaration.Type.IsVar)
-            {
-                typeSymbol = semanticModel.GetTypeInfo(variableDeclaration.Type).Type;
-            }
-            else if (node is ImplicitObjectCreationExpressionSyntax implicitObjectCreation)
-            {
-                typeSymbol = semanticModel.GetTypeInfo(implicitObjectCreation).Type;
-            }
-
-            if (typeSymbol == null || typeSymbol.TypeKind == TypeKind.Error)
-            {
-                continue;
-            }
-
-            if (!AccessibilityRules.IsExternallyVisibleType(typeSymbol))
-            {
-                rejectReason = "body uses a type that is not visible from an external assembly: "
-                    + typeSymbol.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat)
-                    + " (condition c).";
-                return false;
-            }
-        }
-
-        rejectReason = null;
-        return true;
-    }
-
-    private static bool NeedsPropertyIncrementRewrite(SemanticModel semanticModel, SyntaxNode bodyNode)
-    {
-        foreach (SyntaxNode node in bodyNode.DescendantNodes())
-        {
-            if (TransformWorkerProgram.NameofRules.IsInsideNameofArgument(node))
-            {
-                continue;
-            }
-
-            ExpressionSyntax operand = null;
-            if (node is PostfixUnaryExpressionSyntax postfix
-                && (postfix.IsKind(SyntaxKind.PostIncrementExpression)
-                    || postfix.IsKind(SyntaxKind.PostDecrementExpression)))
-            {
-                operand = postfix.Operand;
-            }
-            else if (node is PrefixUnaryExpressionSyntax prefix
-                && (prefix.IsKind(SyntaxKind.PreIncrementExpression)
-                    || prefix.IsKind(SyntaxKind.PreDecrementExpression)))
-            {
-                operand = prefix.Operand;
-            }
-
-            if (operand == null)
-            {
-                continue;
-            }
-
-            ISymbol symbol = semanticModel.GetSymbolInfo(operand).Symbol;
-            if (symbol is IPropertySymbol propertySymbol
-                && (AccessibilityRules.IsInaccessibleAccessor(propertySymbol.GetMethod)
-                    || AccessibilityRules.IsInaccessibleAccessor(propertySymbol.SetMethod)))
-            {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    /// <summary>
-    /// Returns false with null rejectReason when the node is not an inaccessible access site.
-    /// Returns false with a reason when the site is inaccessible but not rewriteable.
-    /// Returns true when the site was registered (or was already present).
-    /// </summary>
-    private static bool TryRegisterInaccessibleAccess(
-        SemanticModel semanticModel,
-        SyntaxNode node,
-        AccessorPlan plan,
-        out string rejectReason)
-    {
-        rejectReason = null;
-
-        if (node is ObjectCreationExpressionSyntax or ImplicitObjectCreationExpressionSyntax)
-        {
-            return TryRegisterObjectCreation(semanticModel, node, out rejectReason);
-        }
-
-        if (node is AssignmentExpressionSyntax assignment
-            && assignment.Parent is InitializerExpressionSyntax)
-        {
-            return TryRegisterInitializerAssignment(semanticModel, assignment, out rejectReason);
-        }
-
-        if (node is InvocationExpressionSyntax invocation)
-        {
-            return TryRegisterInvocation(semanticModel, invocation, plan, out rejectReason);
-        }
-
-        if (node is ElementAccessExpressionSyntax elementAccess)
-        {
-            return TryRegisterElementAccess(semanticModel, elementAccess, out rejectReason);
-        }
-
-        if (node is MemberBindingExpressionSyntax memberBinding)
-        {
-            return TryRegisterMemberBinding(semanticModel, memberBinding, out rejectReason);
-        }
-
-        if (node is AssignmentExpressionSyntax propertyOrFieldAssignment)
-        {
-            return TryRegisterAssignment(
-                semanticModel,
-                propertyOrFieldAssignment,
-                plan,
-                out rejectReason);
-        }
-
-        if (node is MemberAccessExpressionSyntax memberAccess)
-        {
-            return TryRegisterMemberAccess(semanticModel, memberAccess, plan, out rejectReason);
-        }
-
-        if (node is IdentifierNameSyntax or GenericNameSyntax)
-        {
-            return TryRegisterSimpleName(semanticModel, (SimpleNameSyntax)node, plan, out rejectReason);
-        }
-
-        return false;
-    }
-
-    private static bool TryRegisterObjectCreation(
-        SemanticModel semanticModel,
-        SyntaxNode node,
-        out string rejectReason)
-    {
-        rejectReason = null;
-        ISymbol ctorSymbol = semanticModel.GetSymbolInfo(node).Symbol;
-        if (ctorSymbol != null && AccessibilityRules.IsInaccessibleFromExternalAssembly(ctorSymbol))
-        {
-            rejectReason =
-                "inaccessible constructor call has no accessor rewrite shape.";
-            return false;
-        }
-
-        return false;
-    }
-
-    private static bool TryRegisterInitializerAssignment(
-        SemanticModel semanticModel,
-        AssignmentExpressionSyntax assignment,
-        out string rejectReason)
-    {
-        rejectReason = null;
-        // Initializer assignments are always writes (including ImplicitElementAccess indexers).
-        ISymbol initializerSymbol = semanticModel.GetSymbolInfo(assignment.Left).Symbol;
-        bool inaccessibleWrite = initializerSymbol is IPropertySymbol initializerProperty
-            ? AccessibilityRules.IsInaccessibleAccessor(initializerProperty.SetMethod)
-            : initializerSymbol != null
-                && AccessibilityRules.IsInaccessibleFromExternalAssembly(initializerSymbol);
-        if (inaccessibleWrite)
-        {
-            rejectReason =
-                "inaccessible member assignment in an object/collection initializer has no "
-                + "accessor rewrite shape.";
-            return false;
-        }
-
-        return false;
-    }
-
-    private static bool TryRegisterElementAccess(
-        SemanticModel semanticModel,
-        ElementAccessExpressionSyntax elementAccess,
-        out string rejectReason)
-    {
-        rejectReason = null;
-        // Assignment-left ElementAccess is owned by the assignment branch (write context).
-        if (elementAccess.Parent is AssignmentExpressionSyntax parentElementAssignment
-            && parentElementAssignment.Left == elementAccess)
-        {
-            return false;
-        }
-
-        ISymbol symbol = semanticModel.GetSymbolInfo(elementAccess).Symbol;
-        if (symbol is IPropertySymbol indexer && indexer.IsIndexer)
-        {
-            // Standalone ElementAccess is a read — only the getter matters.
-            if (AccessibilityRules.IsInaccessibleAccessor(indexer.GetMethod))
-            {
-                rejectReason =
-                    "inaccessible indexer access has no accessor rewrite shape.";
-                return false;
-            }
-        }
-        else if (symbol != null && AccessibilityRules.IsInaccessibleFromExternalAssembly(symbol))
-        {
-            rejectReason = "inaccessible indexer access has no accessor rewrite shape.";
-            return false;
-        }
-
-        return false;
-    }
-
-    private static bool TryRegisterMemberBinding(
-        SemanticModel semanticModel,
-        MemberBindingExpressionSyntax memberBinding,
-        out string rejectReason)
-    {
-        rejectReason = null;
-        ISymbol bound = semanticModel.GetSymbolInfo(memberBinding.Name).Symbol;
-        if (bound != null
-            && bound is not INamespaceSymbol
-            && bound is not ITypeSymbol
-            && IsInaccessibleBindingTarget(bound))
-        {
-            rejectReason =
-                "inaccessible member access via conditional access has no rewrite shape.";
-            return false;
-        }
-
-        return false;
-    }
-
-    private static bool TryRegisterMemberAccess(
-        SemanticModel semanticModel,
-        MemberAccessExpressionSyntax memberAccess,
-        AccessorPlan plan,
-        out string rejectReason)
-    {
-        rejectReason = null;
-        // Method-group invocation targets are owned by the invocation branch; delegate-typed
-        // field invokes (`this._cb()` / `other._cb()`) register as field reads here.
-        if (memberAccess.Parent is InvocationExpressionSyntax parentInvocation
-            && parentInvocation.Expression == memberAccess)
-        {
-            ISymbol invocationTarget = semanticModel.GetSymbolInfo(memberAccess).Symbol
-                ?? semanticModel.GetSymbolInfo(memberAccess.Name).Symbol;
-            if (invocationTarget is IMethodSymbol)
-            {
-                return false;
-            }
-        }
-
-        if (memberAccess.Parent is AssignmentExpressionSyntax parentAssignment
-            && parentAssignment.Left == memberAccess)
-        {
-            return false;
-        }
-
-        return TryRegisterPropertyOrFieldRead(
-            semanticModel.GetSymbolInfo(memberAccess).Symbol
-            ?? semanticModel.GetSymbolInfo(memberAccess.Name).Symbol,
-            plan,
-            out rejectReason);
-    }
-
-    private static bool TryRegisterSimpleName(
-        SemanticModel semanticModel,
-        SimpleNameSyntax name,
-        AccessorPlan plan,
-        out string rejectReason)
-    {
-        rejectReason = null;
-        if (IsNameHandledByParent(name))
-        {
-            return false;
-        }
-
-        if (name.Parent is AssignmentExpressionSyntax parentAssignment
-            && parentAssignment.Left == name)
-        {
-            return false;
-        }
-
-        // Method-group invocation targets are owned by the invocation branch; delegate-typed
-        // field invokes (`_cb()`) register as field reads here.
-        if (name.Parent is InvocationExpressionSyntax parentInvocation
-            && parentInvocation.Expression == name)
-        {
-            ISymbol invocationTarget = semanticModel.GetSymbolInfo(name).Symbol;
-            if (invocationTarget is IMethodSymbol)
-            {
-                return false;
-            }
-        }
-
-        return TryRegisterPropertyOrFieldRead(
-            semanticModel.GetSymbolInfo(name).Symbol,
-            plan,
-            out rejectReason);
-    }
-
-    private static bool IsInaccessibleBindingTarget(ISymbol bound)
-    {
-        // Member binding only appears in read/invoke contexts (x?.P = v is not valid C#).
-        if (bound is IPropertySymbol propertySymbol)
-        {
-            return AccessibilityRules.IsInaccessibleAccessor(propertySymbol.GetMethod);
-        }
-
-        return AccessibilityRules.IsInaccessibleFromExternalAssembly(bound);
-    }
-
-    private static bool TryRegisterAssignment(
-        SemanticModel semanticModel,
-        AssignmentExpressionSyntax assignment,
-        AccessorPlan plan,
-        out string rejectReason)
-    {
-        rejectReason = null;
-        ISymbol leftSymbol = semanticModel.GetSymbolInfo(assignment.Left).Symbol;
-        if (leftSymbol is IFieldSymbol fieldSymbol)
-        {
-            if (!AccessibilityRules.IsInaccessibleFromExternalAssembly(fieldSymbol))
-            {
-                return false;
-            }
-
-            if (fieldSymbol.IsConst)
-            {
-                return true;
-            }
-
-            plan.GetOrAddField(fieldSymbol);
-            return true;
-        }
-
-        if (leftSymbol is IPropertySymbol propertySymbol)
-        {
-            return TryRegisterPropertyWrite(
-                semanticModel,
-                assignment,
-                propertySymbol,
-                plan,
-                out rejectReason);
-        }
-
-        if (leftSymbol is IEventSymbol)
-        {
-            rejectReason = "inaccessible event add/remove is out of scope for accessor rewrite.";
-            return false;
-        }
-
-        return false;
-    }
-
-    private static bool TryRegisterInvocation(
-        SemanticModel semanticModel,
-        InvocationExpressionSyntax invocation,
-        AccessorPlan plan,
-        out string rejectReason)
-    {
-        rejectReason = null;
-        if (TransformWorkerProgram.NameofRules.IsNameofInvocation(invocation))
-        {
-            return false;
-        }
-
-        ISymbol symbol = semanticModel.GetSymbolInfo(invocation).Symbol;
-        if (symbol is not IMethodSymbol methodSymbol)
-        {
-            return false;
-        }
-
-        if (methodSymbol.MethodKind != MethodKind.Ordinary)
-        {
-            return false;
-        }
-
-        if (!AccessibilityRules.IsInaccessibleFromExternalAssembly(methodSymbol))
-        {
-            return false;
-        }
-
-        if (methodSymbol.IsExtensionMethod)
-        {
-            rejectReason = "inaccessible extension method calls are not rewritten.";
-            return false;
-        }
-
-        if (methodSymbol.IsGenericMethod)
-        {
-            rejectReason = "inaccessible generic method calls are not rewritten.";
-            return false;
-        }
-
-        if (methodSymbol.ReturnsByRef || methodSymbol.ReturnsByRefReadonly)
-        {
-            rejectReason =
-                "inaccessible methods that return by ref have no accessor rewrite shape.";
-            return false;
-        }
-
-        foreach (IParameterSymbol parameter in methodSymbol.Parameters)
-        {
-            if (parameter.RefKind != RefKind.None)
-            {
-                rejectReason =
-                    "inaccessible method calls with ref/out/in parameters are not rewritten.";
-                return false;
-            }
-        }
-
-        foreach (ArgumentSyntax argument in invocation.ArgumentList.Arguments)
-        {
-            if (argument.NameColon != null)
-            {
-                rejectReason =
-                    "inaccessible method calls with named arguments are not rewritten.";
-                return false;
-            }
-        }
-
-        if (invocation.ArgumentList.Arguments.Count != methodSymbol.Parameters.Length)
-        {
-            rejectReason =
-                "inaccessible method calls with omitted optional or expanded params arguments "
-                + "are not rewritten.";
-            return false;
-        }
-
-        plan.GetOrAddMethod(methodSymbol);
-        return true;
-    }
-
-    private static bool TryRegisterPropertyOrFieldRead(
-        ISymbol symbol,
-        AccessorPlan plan,
-        out string rejectReason)
-    {
-        rejectReason = null;
-        if (symbol is IFieldSymbol fieldSymbol)
-        {
-            if (!AccessibilityRules.IsInaccessibleFromExternalAssembly(fieldSymbol))
-            {
-                return false;
-            }
-
-            if (fieldSymbol.IsConst)
-            {
-                return true;
-            }
-
-            plan.GetOrAddField(fieldSymbol);
-            return true;
-        }
-
-        if (symbol is IPropertySymbol propertySymbol)
-        {
-            if (!AccessibilityRules.IsInaccessibleAccessor(propertySymbol.GetMethod))
-            {
-                return false;
-            }
-
-            return TryRegisterPropertyRead(propertySymbol, plan, out rejectReason);
-        }
-
-        if (symbol is IEventSymbol)
-        {
-            rejectReason = "inaccessible event add/remove is out of scope for accessor rewrite.";
-            return false;
-        }
-
-        if (symbol is IMethodSymbol methodSymbol
-            && AccessibilityRules.IsInaccessibleFromExternalAssembly(methodSymbol))
-        {
-            rejectReason =
-                "inaccessible method group (non-invocation) has no accessor rewrite shape.";
-            return false;
-        }
-
-        if (symbol != null
-            && AccessibilityRules.IsInaccessibleFromExternalAssembly(symbol)
-            && symbol is not INamespaceSymbol
-            && symbol is not ITypeSymbol
-            && symbol is not ILocalSymbol
-            && symbol is not IParameterSymbol)
-        {
-            rejectReason = "inaccessible member kind is not field/method/property access.";
-            return false;
-        }
-
-        return false;
-    }
-
-    private static bool TryRegisterPropertyRead(
-        IPropertySymbol propertySymbol,
-        AccessorPlan plan,
-        out string rejectReason)
-    {
-        rejectReason = null;
-        if (propertySymbol.IsIndexer)
-        {
-            rejectReason = "inaccessible indexer access has no accessor rewrite shape.";
-            return false;
-        }
-
-        if (propertySymbol.IsStatic)
-        {
-            rejectReason =
-                "inaccessible static property access has no accessor rewrite shape.";
-            return false;
-        }
-
-        if (propertySymbol.ReturnsByRef || propertySymbol.ReturnsByRefReadonly)
-        {
-            rejectReason =
-                "inaccessible ref-returning properties have no accessor rewrite shape.";
-            return false;
-        }
-
-        plan.GetOrAddPropertyGetter(propertySymbol);
-        return true;
-    }
-
-    private static bool TryRegisterPropertyWrite(
-        SemanticModel semanticModel,
-        AssignmentExpressionSyntax assignment,
-        IPropertySymbol propertySymbol,
-        AccessorPlan plan,
-        out string rejectReason)
-    {
-        rejectReason = null;
-        bool needsGetter = !assignment.IsKind(SyntaxKind.SimpleAssignmentExpression);
-
-        // Why accessibility first: shape gates (indexer/static/ref-return) must not reject fully
-        // public writes such as dict[key]=value or Time.timeScale=0f. The read-side path already
-        // pre-filters with IsInaccessibleAccessor/IsInaccessibleFromExternalAssembly before shape
-        // checks — keep that order here for symmetry.
-        bool setterInaccessible = AccessibilityRules.IsInaccessibleAccessor(propertySymbol.SetMethod);
-        bool getterInaccessible = needsGetter
-            && AccessibilityRules.IsInaccessibleAccessor(propertySymbol.GetMethod);
-        if (!setterInaccessible && !getterInaccessible)
-        {
-            return false;
-        }
-
-        string shapeRejectReason = TryGetPropertyWriteShapeRejectReason(
-            semanticModel,
-            assignment,
-            propertySymbol,
-            needsGetter,
-            setterInaccessible,
-            getterInaccessible);
-        if (shapeRejectReason != null)
-        {
-            rejectReason = shapeRejectReason;
-            return false;
-        }
-
-        if (setterInaccessible)
-        {
-            if (propertySymbol.SetMethod == null)
-            {
-                rejectReason = "inaccessible property has no setter to bind.";
-                return false;
-            }
-
-            plan.GetOrAddPropertySetter(propertySymbol);
-        }
-
-        if (getterInaccessible)
-        {
-            if (propertySymbol.GetMethod == null)
-            {
-                rejectReason = "inaccessible property has no getter to bind.";
-                return false;
-            }
-
-            plan.GetOrAddPropertyGetter(propertySymbol);
-        }
-
-        return true;
-    }
-
-    private static string TryGetPropertyWriteShapeRejectReason(
-        SemanticModel semanticModel,
-        AssignmentExpressionSyntax assignment,
-        IPropertySymbol propertySymbol,
-        bool needsGetter,
-        bool setterInaccessible,
-        bool getterInaccessible)
-    {
-        if (assignment.IsKind(SyntaxKind.CoalesceAssignmentExpression))
-        {
-            return
-                "null-coalescing assignment writes conditionally and has no accessor rewrite shape.";
-        }
-
-        if (needsGetter && !IsSupportedCompoundAssignmentKind(assignment.Kind()))
-        {
-            return
-                "unsupported compound assignment kind has no accessor rewrite shape.";
-        }
-
-        // Compound assignment with a private getter and a public setter has no rewrite shape:
-        // RewritePropertyAssignment only fires when the setter is inaccessible.
-        if (getterInaccessible && !setterInaccessible)
-        {
-            return
-                "compound assignment reading an inaccessible getter with an accessible setter "
-                + "has no accessor rewrite shape.";
-        }
-
-        // Setter delegates are void — consuming the assignment expression value cannot compile.
-        if (assignment.Parent is not ExpressionStatementSyntax)
-        {
-            return
-                "assignment value is consumed; the setter delegate returns void.";
-        }
-
-        // Compound/get+set rewrite embeds the receiver twice; reject side-effecting receivers.
-        if (needsGetter && !IsSideEffectFreeAssignmentReceiver(semanticModel, assignment.Left))
-        {
-            return
-                "receiver with possible side effects would be evaluated twice.";
-        }
-
-        if (propertySymbol.IsIndexer)
-        {
-            return "inaccessible indexer access has no accessor rewrite shape.";
-        }
-
-        if (propertySymbol.IsStatic)
-        {
-            return
-                "inaccessible static property access has no accessor rewrite shape.";
-        }
-
-        if (propertySymbol.ReturnsByRef || propertySymbol.ReturnsByRefReadonly)
-        {
-            return
-                "inaccessible ref-returning properties have no accessor rewrite shape.";
-        }
-
-        return null;
-    }
-
-    internal static bool IsSupportedCompoundAssignmentKind(SyntaxKind kind)
-    {
-        return kind == SyntaxKind.AddAssignmentExpression
-            || kind == SyntaxKind.SubtractAssignmentExpression
-            || kind == SyntaxKind.MultiplyAssignmentExpression
-            || kind == SyntaxKind.DivideAssignmentExpression
-            || kind == SyntaxKind.ModuloAssignmentExpression
-            || kind == SyntaxKind.AndAssignmentExpression
-            || kind == SyntaxKind.ExclusiveOrAssignmentExpression
-            || kind == SyntaxKind.OrAssignmentExpression
-            || kind == SyntaxKind.LeftShiftAssignmentExpression
-            || kind == SyntaxKind.RightShiftAssignmentExpression
-            || kind == SyntaxKind.UnsignedRightShiftAssignmentExpression;
-    }
-
-    /// <summary>
-    /// What: whether an assignment left's receiver chain is free of re-evaluable members
-    /// (properties/methods). Only this/locals/parameters/fields (and type/namespace qualifiers)
-    /// are allowed — FieldRef re-reads the same storage, so field links are idempotent.
-    /// </summary>
-    internal static bool IsSideEffectFreeAssignmentReceiver(
-        SemanticModel semanticModel,
-        ExpressionSyntax left)
-    {
-        ExpressionSyntax receiver = left is MemberAccessExpressionSyntax memberAccess
-            ? memberAccess.Expression
-            : null;
-        if (receiver == null)
-        {
-            // Bare member — implicit this.
-            return true;
-        }
-
-        ExpressionSyntax current = receiver;
-        while (current is MemberAccessExpressionSyntax nested)
-        {
-            ISymbol linkSymbol = semanticModel.GetSymbolInfo(nested.Name).Symbol
-                ?? semanticModel.GetSymbolInfo(nested).Symbol;
-            if (!IsSideEffectFreeReceiverLink(linkSymbol))
-            {
-                return false;
-            }
-
-            current = nested.Expression;
-        }
-
-        if (current is ThisExpressionSyntax || current is BaseExpressionSyntax)
-        {
-            return true;
-        }
-
-        if (current is IdentifierNameSyntax)
-        {
-            ISymbol headSymbol = semanticModel.GetSymbolInfo(current).Symbol;
-            return headSymbol is ILocalSymbol
-                || headSymbol is IParameterSymbol
-                || headSymbol is IFieldSymbol
-                || headSymbol is ITypeSymbol
-                || headSymbol is INamespaceSymbol;
-        }
-
-        return false;
-    }
-
-    private static bool IsSideEffectFreeReceiverLink(ISymbol linkSymbol)
-    {
-        // Fields re-read the same storage; type/namespace qualifiers are not evaluated.
-        return linkSymbol is IFieldSymbol
-            || linkSymbol is ITypeSymbol
-            || linkSymbol is INamespaceSymbol;
-    }
-
-    public static bool IsNameHandledByParent(SimpleNameSyntax node)
-    {
-        if (node.Parent is MemberAccessExpressionSyntax memberAccess && memberAccess.Name == node)
-        {
-            return true;
-        }
-
-        if (node.Parent is QualifiedNameSyntax qualifiedName && qualifiedName.Right == node)
-        {
-            return true;
-        }
-
-        if (node.Parent is MemberBindingExpressionSyntax memberBinding && memberBinding.Name == node)
-        {
-            return true;
-        }
-
-        // Invocation targets are NOT handled here: method groups are skipped by the caller after
-        // a symbol check; delegate-typed field invokes must reach the field-read path.
-        return false;
-    }
-}
-
-/// <summary>
-/// Qualifies bare instance/static member references and, when an accessor plan is supplied,
-/// rewrites inaccessible field/method/property accesses into Harmony accessor delegate calls.
-/// </summary>
-internal sealed class ShimBodyRewriter : CSharpSyntaxRewriter
-{
-    private readonly SemanticModel _semanticModel;
-    private readonly INamedTypeSymbol _targetType;
-    private readonly AccessorPlan _accessorPlan;
-    private readonly AddedMethodCatalog _addedMethodCatalog;
-    private readonly AddedFieldCatalog _addedFieldCatalog;
-
-    public ShimBodyRewriter(
-        SemanticModel semanticModel,
-        INamedTypeSymbol targetType,
-        AccessorPlan accessorPlan,
-        AddedMethodCatalog addedMethodCatalog,
-        AddedFieldCatalog addedFieldCatalog)
-    {
-        _semanticModel = semanticModel;
-        _targetType = targetType;
-        _accessorPlan = accessorPlan;
-        _addedMethodCatalog = addedMethodCatalog ?? new AddedMethodCatalog();
-        _addedFieldCatalog = addedFieldCatalog ?? new AddedFieldCatalog();
-    }
-
-    public override SyntaxNode VisitThisExpression(ThisExpressionSyntax node)
-    {
-        return SyntaxFactory.IdentifierName(TransformWorkerProgramMarker.InstanceParameterName)
-            .WithTriviaFrom(node);
-    }
-
-    public override SyntaxNode VisitIdentifierName(IdentifierNameSyntax node)
-    {
-        return VisitName(node, node);
-    }
-
-    public override SyntaxNode VisitGenericName(GenericNameSyntax node)
-    {
-        return VisitName(node, node);
-    }
-
-    public override SyntaxNode VisitInterpolation(InterpolationSyntax node)
-    {
-        InterpolationSyntax visited = (InterpolationSyntax)base.VisitInterpolation(node);
-
-        // Why: a top-level ':' in an interpolation hole starts a format clause, so a
-        // rewrite that inserts bare `global::` yields CS0103 ('global'). Parenthesizing
-        // keeps the alias qualifier out of the format-clause scan and still coexists
-        // with alignment/format clauses. Nested positions do not need parentheses, but
-        // wrapping whenever an AliasQualifiedNameSyntax is present is always safe.
-        // Alignment widths are integer expressions and hit the same ':' scan, so they
-        // need the same wrapping; format clauses are literal text and need none.
-        ExpressionSyntax parenthesizedExpression = ParenthesizeIfAliasQualified(visited.Expression);
-        if (!ReferenceEquals(parenthesizedExpression, visited.Expression))
-        {
-            visited = visited.WithExpression(parenthesizedExpression);
-        }
-
-        InterpolationAlignmentClauseSyntax alignmentClause = visited.AlignmentClause;
-        if (alignmentClause != null)
-        {
-            ExpressionSyntax parenthesizedAlignment = ParenthesizeIfAliasQualified(alignmentClause.Value);
-            if (!ReferenceEquals(parenthesizedAlignment, alignmentClause.Value))
-            {
-                visited = visited.WithAlignmentClause(
-                    alignmentClause.WithValue(parenthesizedAlignment));
-            }
-        }
-
-        return visited;
-    }
-
-    private static ExpressionSyntax ParenthesizeIfAliasQualified(ExpressionSyntax expression)
-    {
-        if (expression is ParenthesizedExpressionSyntax)
-        {
-            return expression;
-        }
-
-        foreach (SyntaxNode descendant in expression.DescendantNodesAndSelf())
-        {
-            if (descendant is AliasQualifiedNameSyntax)
-            {
-                return SyntaxFactory.ParenthesizedExpression(expression);
-            }
-        }
-
-        return expression;
-    }
-
-    public override SyntaxNode VisitInvocationExpression(InvocationExpressionSyntax node)
-    {
-        // Why first: added-member rewrite must not depend on _accessorPlan. Transplant bodies
-        // have a null plan and would skip rewrite; delegation bodies would otherwise bind
-        // Harmony accessors onto members that do not exist on the compiled type (B1).
-        if (TransformWorkerProgram.NameofRules.IsNameofInvocation(node))
-        {
-            ExpressionSyntax folded = TryFoldNameofAddedMember(node);
-            if (folded != null)
-            {
-                return folded;
-            }
-
-            return base.VisitInvocationExpression(node);
-        }
-
-        if (TransformWorkerProgram.NameofRules.IsInsideNameofArgument(node))
-        {
-            return base.VisitInvocationExpression(node);
-        }
-
-        ISymbol invokedSymbol = _semanticModel.GetSymbolInfo(node).Symbol;
-        if (TransformWorkerProgram.IsConditionalAccessReceiverSpine(node))
-        {
-            // Why not rewrite the spine invocation: ExtractReceiver cannot recover a
-            // MemberBinding/ElementBinding receiver and would emit a parse-invalid shim.
-            // Arguments and lambdas are not on the spine; base.Visit still rewrites those.
-            return base.VisitInvocationExpression(node);
-        }
-
-        if (invokedSymbol is IMethodSymbol addedMethod
-            && addedMethod.MethodKind == MethodKind.Ordinary)
-        {
-            AddedMethodBinding binding = _addedMethodCatalog.FindOrNull(BuildAddedMethodKey(addedMethod));
-            if (binding != null)
-            {
-                return RewriteAddedMethodInvocation(node, addedMethod, binding);
-            }
-        }
-
-        if (_accessorPlan == null)
-        {
-            return base.VisitInvocationExpression(node);
-        }
-
-        ISymbol symbol = invokedSymbol;
-        if (symbol is not IMethodSymbol methodSymbol
-            || methodSymbol.MethodKind != MethodKind.Ordinary
-            || !AccessibilityRules.IsInaccessibleFromExternalAssembly(methodSymbol)
-            || methodSymbol.IsExtensionMethod)
-        {
-            return base.VisitInvocationExpression(node);
-        }
-
-        AccessorEntry entry = _accessorPlan.GetOrAddMethod(methodSymbol);
-        List<ArgumentSyntax> arguments = new List<ArgumentSyntax>();
-        if (!methodSymbol.IsStatic)
-        {
-            ExpressionSyntax receiver = ExtractReceiver(node.Expression);
-            arguments.Add(SyntaxFactory.Argument(VisitReceiver(receiver)));
-        }
-
-        foreach (ArgumentSyntax argument in node.ArgumentList.Arguments)
-        {
-            arguments.Add((ArgumentSyntax)Visit(argument));
-        }
-
-        return SyntaxFactory.InvocationExpression(
-                SyntaxFactory.IdentifierName(entry.DelegateFieldName),
-                SyntaxFactory.ArgumentList(SyntaxFactory.SeparatedList(arguments)))
-            .WithTriviaFrom(node);
-    }
-
-    private ExpressionSyntax TryFoldNameofAddedMember(InvocationExpressionSyntax nameofInvocation)
-    {
-        if (nameofInvocation.ArgumentList.Arguments.Count != 1)
-        {
-            return null;
-        }
-
-        ExpressionSyntax argument = nameofInvocation.ArgumentList.Arguments[0].Expression;
-        ISymbol symbol = ResolveNameofArgumentSymbol(argument);
-        if (symbol is IMethodSymbol methodSymbol
-            && _addedMethodCatalog.Contains(BuildAddedMethodKey(methodSymbol)))
-        {
-            return SyntaxFactory.LiteralExpression(
-                    SyntaxKind.StringLiteralExpression,
-                    SyntaxFactory.Literal(methodSymbol.Name))
-                .WithTriviaFrom(nameofInvocation);
-        }
-
-        if (symbol is IFieldSymbol fieldSymbol
-            && _addedFieldCatalog.FindOrNull(
-                TransformWorkerProgram.FormatAddedFieldKeyFromSymbol(fieldSymbol)) != null)
-        {
-            return SyntaxFactory.LiteralExpression(
-                    SyntaxKind.StringLiteralExpression,
-                    SyntaxFactory.Literal(fieldSymbol.Name))
-                .WithTriviaFrom(nameofInvocation);
-        }
-
-        return null;
-    }
-
-    private ISymbol ResolveNameofArgumentSymbol(ExpressionSyntax argument)
-    {
-        // Why CandidateSymbols: nameof(method) is a method group, so Symbol is often null
-        // and the unique candidate is the added method we need to fold.
-        SymbolInfo symbolInfo = _semanticModel.GetSymbolInfo(argument);
-        if (symbolInfo.Symbol != null)
-        {
-            return symbolInfo.Symbol;
-        }
-
-        if (symbolInfo.CandidateSymbols.Length == 1)
-        {
-            return symbolInfo.CandidateSymbols[0];
-        }
-
-        return null;
-    }
-
-    private SyntaxNode RewriteAddedMethodInvocation(
-        InvocationExpressionSyntax node,
-        IMethodSymbol addedMethod,
-        AddedMethodBinding binding)
-    {
-        string qualifiedShimType = string.IsNullOrEmpty(binding.NamespaceName)
-            ? "global::" + binding.ShimTypeName
-            : "global::" + binding.NamespaceName + "." + binding.ShimTypeName;
-        ExpressionSyntax shimTypeExpression = SyntaxFactory.ParseTypeName(qualifiedShimType);
-        ExpressionSyntax shimAccess = SyntaxFactory.MemberAccessExpression(
-            SyntaxKind.SimpleMemberAccessExpression,
-            shimTypeExpression,
-            SyntaxFactory.IdentifierName(binding.ShimMethodName));
-
-        List<ArgumentSyntax> arguments = new List<ArgumentSyntax>();
-        if (!addedMethod.IsStatic)
-        {
-            ExpressionSyntax receiver = ExtractReceiver(node.Expression);
-            arguments.Add(SyntaxFactory.Argument(VisitReceiver(receiver)));
-        }
-
-        foreach (ArgumentSyntax argument in node.ArgumentList.Arguments)
-        {
-            arguments.Add((ArgumentSyntax)Visit(argument));
-        }
-
-        return SyntaxFactory.InvocationExpression(
-                shimAccess,
-                SyntaxFactory.ArgumentList(SyntaxFactory.SeparatedList(arguments)))
-            .WithTriviaFrom(node);
-    }
-
-    private static string BuildAddedMethodKey(IMethodSymbol methodSymbol)
-    {
-        string[] parameterTypeFullNames = methodSymbol.Parameters
-            .Select(CecilTypeNames.ToParameterTypeFullName)
-            .ToArray();
-        return methodSymbol.ContainingType == null
-            ? methodSymbol.Name
-            : CecilTypeNames.ToMetadataName(methodSymbol.ContainingType)
-                + "::" + methodSymbol.Name + "("
-                + string.Join(",", parameterTypeFullNames) + ")";
-    }
-
-    public override SyntaxNode VisitAssignmentExpression(AssignmentExpressionSyntax node)
-    {
-        // Object/collection initializer member names must stay bare identifiers.
-        if (node.Parent is InitializerExpressionSyntax)
-        {
-            return base.VisitAssignmentExpression(node);
-        }
-
-        if (TransformWorkerProgram.NameofRules.IsInsideNameofArgument(node))
-        {
-            return base.VisitAssignmentExpression(node);
-        }
-
-        AddedFieldBinding assignedField = FindStoreBinding(_semanticModel.GetSymbolInfo(node.Left).Symbol);
-        if (assignedField != null)
-        {
-            return RewriteAddedFieldAssignment(node, assignedField);
-        }
-
-        if (_accessorPlan == null)
-        {
-            return base.VisitAssignmentExpression(node);
-        }
-
-        ISymbol leftSymbol = _semanticModel.GetSymbolInfo(node.Left).Symbol;
-        if (leftSymbol is IPropertySymbol propertySymbol
-            && !propertySymbol.IsIndexer
-            && !propertySymbol.IsStatic
-            && AccessibilityRules.IsInaccessibleAccessor(propertySymbol.SetMethod))
-        {
-            return RewritePropertyAssignment(node, propertySymbol);
-        }
-
-        if (leftSymbol is IFieldSymbol fieldSymbol
-            && !fieldSymbol.IsConst
-            && AccessibilityRules.IsInaccessibleFromExternalAssembly(fieldSymbol))
-        {
-            AccessorEntry entry = _accessorPlan.GetOrAddField(fieldSymbol);
-            ExpressionSyntax fieldRefCall = CreateFieldRefInvocation(
-                entry,
-                VisitReceiver(ExtractReceiver(node.Left)));
-            return node
-                .WithLeft(fieldRefCall)
-                .WithRight((ExpressionSyntax)Visit(node.Right))
-                .WithTriviaFrom(node);
-        }
-
-        return base.VisitAssignmentExpression(node);
-    }
-
-    public override SyntaxNode VisitPrefixUnaryExpression(PrefixUnaryExpressionSyntax node)
-    {
-        if (TransformWorkerProgram.IsIncrementOrDecrement(node.Kind()))
-        {
-            AddedFieldBinding binding = FindStoreBinding(_semanticModel.GetSymbolInfo(node.Operand).Symbol);
-            if (binding != null)
-            {
-                return RewriteAddedFieldIncrement(node.Operand, binding, node);
-            }
-        }
-
-        return base.VisitPrefixUnaryExpression(node);
-    }
-
-    public override SyntaxNode VisitPostfixUnaryExpression(PostfixUnaryExpressionSyntax node)
-    {
-        if (TransformWorkerProgram.IsIncrementOrDecrement(node.Kind()))
-        {
-            AddedFieldBinding binding = FindStoreBinding(_semanticModel.GetSymbolInfo(node.Operand).Symbol);
-            if (binding != null)
-            {
-                return RewriteAddedFieldIncrement(node.Operand, binding, node);
-            }
-        }
-
-        return base.VisitPostfixUnaryExpression(node);
-    }
-
-    private AddedFieldBinding FindStoreBinding(ISymbol symbol)
-    {
-        if (symbol is not IFieldSymbol fieldSymbol)
-        {
-            return null;
-        }
-
-        AddedFieldBinding binding = _addedFieldCatalog.FindOrNull(
-            TransformWorkerProgram.FormatAddedFieldKeyFromSymbol(fieldSymbol));
-        if (binding == null || !binding.IsStoreRewriteable)
-        {
-            return null;
-        }
-
-        return binding;
-    }
-
-    private AddedFieldBinding FindAnyAddedBinding(ISymbol symbol)
-    {
-        if (symbol is not IFieldSymbol fieldSymbol)
-        {
-            return null;
-        }
-
-        return _addedFieldCatalog.FindOrNull(
-            TransformWorkerProgram.FormatAddedFieldKeyFromSymbol(fieldSymbol));
-    }
-
-    private SyntaxNode TryRewriteAddedFieldRead(
-        ISymbol symbol,
-        ExpressionSyntax receiverSyntax,
-        SyntaxNode triviaSource)
-    {
-        AddedFieldBinding binding = FindAnyAddedBinding(symbol);
-        if (binding == null || binding.UnavailableReason != null)
-        {
-            return null;
-        }
-
-        if (binding.IsConst)
-        {
-            ExpressionSyntax literal = TransformWorkerProgram.TryCreateConstantLiteral(
-                binding.ConstantValue,
-                binding.FieldType);
-            if (literal == null)
-            {
-                return null;
-            }
-
-            _addedFieldCatalog.MarkConstFold(binding.FieldKey);
-            return literal.WithTriviaFrom(triviaSource);
-        }
-
-        if (!binding.IsStoreRewriteable)
-        {
-            return null;
-        }
-
-        return CreateAddedFieldGetOrInit(binding, receiverSyntax).WithTriviaFrom(triviaSource);
-    }
-
-    private SyntaxNode RewriteAddedFieldAssignment(
-        AssignmentExpressionSyntax node,
-        AddedFieldBinding binding)
-    {
-        ExpressionSyntax receiver = ExtractAddedFieldReceiver(node.Left, binding.IsStatic);
-        ExpressionSyntax visitedRight = (ExpressionSyntax)Visit(node.Right);
-        if (node.IsKind(SyntaxKind.SimpleAssignmentExpression))
-        {
-            return CreateAddedFieldSet(binding, receiver, visitedRight).WithTriviaFrom(node);
-        }
-
-        SyntaxKind binaryKind = GetCompoundAssignmentBinaryKind(node.Kind());
-        ExpressionSyntax getCall = CreateAddedFieldGetOrInit(binding, receiver);
-        ExpressionSyntax combined = SyntaxFactory.BinaryExpression(binaryKind, getCall, visitedRight);
-        return CreateAddedFieldSet(
-                binding,
-                receiver,
-                CastToAddedFieldType(combined, binding.FieldType))
-            .WithTriviaFrom(node);
-    }
-
-    private SyntaxNode RewriteAddedFieldIncrement(
-        ExpressionSyntax operand,
-        AddedFieldBinding binding,
-        SyntaxNode triviaSource)
-    {
-        ExpressionSyntax receiver = ExtractAddedFieldReceiver(operand, binding.IsStatic);
-        ExpressionSyntax getCall = CreateAddedFieldGetOrInit(binding, receiver);
-        SyntaxKind binaryKind = IsDecrementNode(triviaSource)
-            ? SyntaxKind.SubtractExpression
-            : SyntaxKind.AddExpression;
-        ExpressionSyntax combined = SyntaxFactory.BinaryExpression(
-            binaryKind,
-            getCall,
-            SyntaxFactory.LiteralExpression(
-                SyntaxKind.NumericLiteralExpression,
-                SyntaxFactory.Literal(1)));
-        return CreateAddedFieldSet(
-                binding,
-                receiver,
-                CastToAddedFieldType(combined, binding.FieldType))
-            .WithTriviaFrom(triviaSource);
-    }
-
-    private static bool IsDecrementNode(SyntaxNode node)
-    {
-        if (node is PrefixUnaryExpressionSyntax prefix)
-        {
-            return prefix.IsKind(SyntaxKind.PreDecrementExpression);
-        }
-
-        return node is PostfixUnaryExpressionSyntax postfix
-            && postfix.IsKind(SyntaxKind.PostDecrementExpression);
-    }
-
-    // Why cast: C# compound assignment and ++/-- apply a conversion back to the field type
-    // (byte += 1 is (byte)(byte + 1)). Emitting the binary without that conversion is CS1503.
-    private static ExpressionSyntax CastToAddedFieldType(ExpressionSyntax expression, ITypeSymbol fieldType)
-    {
-        TypeSyntax typeSyntax = SyntaxFactory.ParseTypeName(
-            fieldType.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat));
-        return SyntaxFactory.CastExpression(
-            typeSyntax,
-            SyntaxFactory.ParenthesizedExpression(expression));
-    }
-
-    private ExpressionSyntax ExtractAddedFieldReceiver(ExpressionSyntax expression, bool isStatic)
-    {
-        if (isStatic)
-        {
-            return null;
-        }
-
-        ExpressionSyntax receiver = ExtractReceiver(expression);
-        if (receiver is ThisExpressionSyntax || receiver is BaseExpressionSyntax)
-        {
-            return SyntaxFactory.IdentifierName(TransformWorkerProgramMarker.InstanceParameterName);
-        }
-
-        return VisitReceiver(receiver);
-    }
-
-    private InvocationExpressionSyntax CreateAddedFieldGetOrInit(
-        AddedFieldBinding binding,
-        ExpressionSyntax receiver)
-    {
-        _addedFieldCatalog.MarkStoreRewrite(binding.FieldKey);
-        string methodName = binding.IsStatic
-            ? TransformWorkerProgramMarker.AddedFieldGetOrInitStaticMethodName
-            : TransformWorkerProgramMarker.AddedFieldGetOrInitMethodName;
-        List<ArgumentSyntax> arguments = new List<ArgumentSyntax>();
-        if (!binding.IsStatic)
-        {
-            arguments.Add(SyntaxFactory.Argument(receiver));
-        }
-
-        arguments.Add(
-            SyntaxFactory.Argument(
-                SyntaxFactory.LiteralExpression(
-                    SyntaxKind.StringLiteralExpression,
-                    SyntaxFactory.Literal(binding.FieldKey))));
-        arguments.Add(SyntaxFactory.Argument(CreateAddedFieldInitializer(binding)));
-        return CreateAddedFieldStoreInvocation(methodName, binding.FieldType, arguments);
-    }
-
-    private InvocationExpressionSyntax CreateAddedFieldSet(
-        AddedFieldBinding binding,
-        ExpressionSyntax receiver,
-        ExpressionSyntax value)
-    {
-        _addedFieldCatalog.MarkStoreRewrite(binding.FieldKey);
-        string methodName = binding.IsStatic
-            ? TransformWorkerProgramMarker.AddedFieldSetStaticMethodName
-            : TransformWorkerProgramMarker.AddedFieldSetMethodName;
-        List<ArgumentSyntax> arguments = new List<ArgumentSyntax>();
-        if (!binding.IsStatic)
-        {
-            arguments.Add(SyntaxFactory.Argument(receiver));
-        }
-
-        arguments.Add(
-            SyntaxFactory.Argument(
-                SyntaxFactory.LiteralExpression(
-                    SyntaxKind.StringLiteralExpression,
-                    SyntaxFactory.Literal(binding.FieldKey))));
-        arguments.Add(SyntaxFactory.Argument(value));
-        return CreateAddedFieldStoreInvocation(methodName, binding.FieldType, arguments);
-    }
-
-    private static ExpressionSyntax CreateAddedFieldInitializer(AddedFieldBinding binding)
-    {
-        if (binding.Initializer == null)
-        {
-            return SyntaxFactory.LiteralExpression(SyntaxKind.NullLiteralExpression);
-        }
-
-        ExpressionSyntax cloned = SyntaxFactory.ParseExpression(binding.Initializer.ToString());
-        return SyntaxFactory.ParenthesizedLambdaExpression(
-                SyntaxFactory.ParameterList(),
-                cloned)
-            .WithModifiers(SyntaxFactory.TokenList(SyntaxFactory.Token(SyntaxKind.StaticKeyword)));
-    }
-
-    private static InvocationExpressionSyntax CreateAddedFieldStoreInvocation(
-        string methodName,
-        ITypeSymbol fieldType,
-        List<ArgumentSyntax> arguments)
-    {
-        TypeSyntax storeType = SyntaxFactory.ParseTypeName(
-            TransformWorkerProgramMarker.AddedFieldStoreTypeName);
-        TypeSyntax typeArgument = SyntaxFactory.ParseTypeName(
-            fieldType.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat));
-        GenericNameSyntax genericName = SyntaxFactory.GenericName(SyntaxFactory.Identifier(methodName))
-            .WithTypeArgumentList(
-                SyntaxFactory.TypeArgumentList(SyntaxFactory.SingletonSeparatedList(typeArgument)));
-        return SyntaxFactory.InvocationExpression(
-            SyntaxFactory.MemberAccessExpression(
-                SyntaxKind.SimpleMemberAccessExpression,
-                storeType,
-                genericName),
-            SyntaxFactory.ArgumentList(SyntaxFactory.SeparatedList(arguments)));
-    }
-
-    private static bool IsAssignmentLeft(SyntaxNode node)
-    {
-        return node.Parent is AssignmentExpressionSyntax assignment && assignment.Left == node;
-    }
-
-    private static bool IsIncrementOperand(SyntaxNode node)
-    {
-        if (node.Parent is PrefixUnaryExpressionSyntax prefix
-            && TransformWorkerProgram.IsIncrementOrDecrement(prefix.Kind()))
-        {
-            return true;
-        }
-
-        return node.Parent is PostfixUnaryExpressionSyntax postfix
-            && TransformWorkerProgram.IsIncrementOrDecrement(postfix.Kind());
-    }
-
-    public override SyntaxNode VisitMemberAccessExpression(MemberAccessExpressionSyntax node)
-    {
-        if (TransformWorkerProgram.NameofRules.IsInsideNameofArgument(node))
-        {
-            return base.VisitMemberAccessExpression(node);
-        }
-
-        ISymbol symbol = _semanticModel.GetSymbolInfo(node).Symbol
-            ?? _semanticModel.GetSymbolInfo(node.Name).Symbol;
-        if (!IsAssignmentLeft(node) && !IsIncrementOperand(node))
-        {
-            SyntaxNode addedFieldRead = TryRewriteAddedFieldRead(symbol, node.Expression, node);
-            if (addedFieldRead != null)
-            {
-                return addedFieldRead;
-            }
-        }
-
-        if (_accessorPlan == null)
-        {
-            return base.VisitMemberAccessExpression(node);
-        }
-
-        // Method-group invocation targets stay with VisitInvocationExpression; field/property
-        // delegate invokes (`this._cb()`) must rewrite here so the call becomes `__F__(recv)()`.
-        if (node.Parent is InvocationExpressionSyntax invocation
-            && invocation.Expression == node
-            && symbol is IMethodSymbol)
-        {
-            return base.VisitMemberAccessExpression(node);
-        }
-
-        if (node.Parent is AssignmentExpressionSyntax assignment && assignment.Left == node)
-        {
-            return base.VisitMemberAccessExpression(node);
-        }
-
-        ExpressionSyntax rewritten = TryRewriteInaccessibleRead(symbol, node.Expression, node);
-        if (rewritten != null)
-        {
-            return rewritten;
-        }
-
-        return base.VisitMemberAccessExpression(node);
-    }
-
-    private SyntaxNode VisitName(SimpleNameSyntax node, SyntaxNode original)
-    {
-        if (IsNameSideOfLargerExpression(node))
-        {
-            return original;
-        }
-
-        ISymbol symbol = _semanticModel.GetSymbolInfo(node).Symbol;
-        if (symbol == null)
-        {
-            return original;
-        }
-
-        SyntaxNode addedFieldRead = TryRewriteUnownedAddedFieldRead(node, symbol);
-        if (addedFieldRead != null)
-        {
-            return addedFieldRead;
-        }
-
-        // Local/anonymous functions are emitted into the shim assembly — keep bare calls.
-        if (IsLocalOrAnonymousFunctionSymbol(symbol))
-        {
-            return original;
-        }
-
-        SyntaxNode accessorRead = TryRewriteNameAsAccessorRead(node, symbol);
-        if (accessorRead != null)
-        {
-            return accessorRead;
-        }
-
-        (bool owned, bool isStatic, INamedTypeSymbol containingType) ownership = ResolveOwnedMember(symbol);
-        if (!ownership.owned)
-        {
-            return original;
-        }
-
-        return QualifyOwnedMemberAccess(node, ownership.isStatic, ownership.containingType);
-    }
-
-    private static bool IsNameSideOfLargerExpression(SimpleNameSyntax node)
-    {
-        return IsMemberAccessNameSide(node)
-            || IsQualifiedNameRightSide(node)
-            || IsMemberBindingName(node)
-            || IsObjectOrCollectionInitializerMemberName(node);
-    }
-
-    private SyntaxNode TryRewriteUnownedAddedFieldRead(SimpleNameSyntax node, ISymbol symbol)
-    {
-        if (IsAssignmentLeft(node)
-            || IsIncrementOperand(node)
-            || TransformWorkerProgram.NameofRules.IsInsideNameofArgument(node))
-        {
-            return null;
-        }
-
-        return TryRewriteAddedFieldRead(
-            symbol,
-            SyntaxFactory.IdentifierName(TransformWorkerProgramMarker.InstanceParameterName),
-            node);
-    }
-
-    private static bool IsLocalOrAnonymousFunctionSymbol(ISymbol symbol)
-    {
-        return symbol is IMethodSymbol methodSymbol
-            && (methodSymbol.MethodKind == MethodKind.LocalFunction
-                || methodSymbol.MethodKind == MethodKind.AnonymousFunction);
-    }
-
-    private SyntaxNode TryRewriteNameAsAccessorRead(SimpleNameSyntax node, ISymbol symbol)
-    {
-        // nameof(...) and assignment left sides must keep a member-reference shape: qualify only,
-        // never rewrite to an accessor read (Func<> call results are not assignable).
-        bool suppressAccessorRead = TransformWorkerProgram.NameofRules.IsInsideNameofArgument(node)
-            || (node.Parent is AssignmentExpressionSyntax assignmentLeft
-                && assignmentLeft.Left == node);
-        if (_accessorPlan == null || suppressAccessorRead)
-        {
-            return null;
-        }
-
-        return TryRewriteInaccessibleRead(
-            symbol,
-            SyntaxFactory.IdentifierName(TransformWorkerProgramMarker.InstanceParameterName),
-            node);
-    }
-
-    private static SyntaxNode QualifyOwnedMemberAccess(
-        SimpleNameSyntax node,
-        bool isStatic,
-        INamedTypeSymbol containingType)
-    {
-        if (isStatic)
-        {
-            TypeSyntax typeSyntax = SyntaxFactory.ParseTypeName(
-                containingType.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat));
-            return SyntaxFactory.MemberAccessExpression(
-                    SyntaxKind.SimpleMemberAccessExpression,
-                    typeSyntax,
-                    (SimpleNameSyntax)node.WithoutTrivia())
-                .WithTriviaFrom(node);
-        }
-
-        return SyntaxFactory.MemberAccessExpression(
-                SyntaxKind.SimpleMemberAccessExpression,
-                SyntaxFactory.IdentifierName(TransformWorkerProgramMarker.InstanceParameterName),
-                (SimpleNameSyntax)node.WithoutTrivia())
-            .WithTriviaFrom(node);
-    }
-
-    private ExpressionSyntax TryRewriteInaccessibleRead(
-        ISymbol symbol,
-        ExpressionSyntax receiverSyntax,
-        SyntaxNode triviaSource)
-    {
-        if (symbol is IFieldSymbol fieldSymbol
-            && !fieldSymbol.IsConst
-            && AccessibilityRules.IsInaccessibleFromExternalAssembly(fieldSymbol))
-        {
-            AccessorEntry entry = _accessorPlan.GetOrAddField(fieldSymbol);
-            return CreateFieldRefInvocation(entry, VisitReceiver(receiverSyntax))
-                .WithTriviaFrom(triviaSource);
-        }
-
-        if (symbol is IPropertySymbol propertySymbol
-            && !propertySymbol.IsIndexer
-            && !propertySymbol.IsStatic
-            && AccessibilityRules.IsInaccessibleAccessor(propertySymbol.GetMethod))
-        {
-            AccessorEntry entry = _accessorPlan.GetOrAddPropertyGetter(propertySymbol);
-            return CreateDelegateInvocation(
-                    entry.DelegateFieldName,
-                    new[] { VisitReceiver(receiverSyntax) })
-                .WithTriviaFrom(triviaSource);
-        }
-
-        return null;
-    }
-
-    private SyntaxNode RewritePropertyAssignment(
-        AssignmentExpressionSyntax node,
-        IPropertySymbol propertySymbol)
-    {
-        ExpressionSyntax receiver = ExtractReceiver(node.Left);
-        ExpressionSyntax visitedReceiver = VisitReceiver(receiver);
-        ExpressionSyntax visitedRight = (ExpressionSyntax)Visit(node.Right);
-        AccessorEntry setter = _accessorPlan.GetOrAddPropertySetter(propertySymbol);
-
-        if (node.IsKind(SyntaxKind.SimpleAssignmentExpression))
-        {
-            return CreateDelegateInvocation(
-                    setter.DelegateFieldName,
-                    new[] { visitedReceiver, visitedRight })
-                .WithTriviaFrom(node);
-        }
-
-        AccessorEntry getter = _accessorPlan.GetOrAddPropertyGetter(propertySymbol);
-        ExpressionSyntax getCall = CreateDelegateInvocation(
-            getter.DelegateFieldName,
-            new[] { visitedReceiver });
-        SyntaxKind binaryKind = GetCompoundAssignmentBinaryKind(node.Kind());
-        ExpressionSyntax combined = SyntaxFactory.BinaryExpression(binaryKind, getCall, visitedRight);
-        return CreateDelegateInvocation(
-                setter.DelegateFieldName,
-                new[] { visitedReceiver, combined })
-            .WithTriviaFrom(node);
-    }
-
-    private static SyntaxKind GetCompoundAssignmentBinaryKind(SyntaxKind assignmentKind)
-    {
-        return assignmentKind switch
-        {
-            SyntaxKind.AddAssignmentExpression => SyntaxKind.AddExpression,
-            SyntaxKind.SubtractAssignmentExpression => SyntaxKind.SubtractExpression,
-            SyntaxKind.MultiplyAssignmentExpression => SyntaxKind.MultiplyExpression,
-            SyntaxKind.DivideAssignmentExpression => SyntaxKind.DivideExpression,
-            SyntaxKind.ModuloAssignmentExpression => SyntaxKind.ModuloExpression,
-            SyntaxKind.AndAssignmentExpression => SyntaxKind.BitwiseAndExpression,
-            SyntaxKind.ExclusiveOrAssignmentExpression => SyntaxKind.ExclusiveOrExpression,
-            SyntaxKind.OrAssignmentExpression => SyntaxKind.BitwiseOrExpression,
-            SyntaxKind.LeftShiftAssignmentExpression => SyntaxKind.LeftShiftExpression,
-            SyntaxKind.RightShiftAssignmentExpression => SyntaxKind.RightShiftExpression,
-            SyntaxKind.UnsignedRightShiftAssignmentExpression => SyntaxKind.UnsignedRightShiftExpression,
-            // Eligibility must reject unsupported compounds (including ??=) before rewrite.
-            _ => throw new System.InvalidOperationException(
-                "Unsupported compound assignment kind reached property rewrite: " + assignmentKind)
-        };
-    }
-
-    private static ExpressionSyntax CreateFieldRefInvocation(
-        AccessorEntry entry,
-        ExpressionSyntax visitedReceiver)
-    {
-        if (entry.FieldSymbol.IsStatic)
-        {
-            return CreateDelegateInvocation(entry.DelegateFieldName, Array.Empty<ExpressionSyntax>());
-        }
-
-        return CreateDelegateInvocation(entry.DelegateFieldName, new[] { visitedReceiver });
-    }
-
-    private static ExpressionSyntax CreateDelegateInvocation(
-        string delegateFieldName,
-        IReadOnlyList<ExpressionSyntax> arguments)
-    {
-        SeparatedSyntaxList<ArgumentSyntax> argumentList = SyntaxFactory.SeparatedList(
-            arguments.Select(SyntaxFactory.Argument));
-        return SyntaxFactory.InvocationExpression(
-            SyntaxFactory.IdentifierName(delegateFieldName),
-            SyntaxFactory.ArgumentList(argumentList));
-    }
-
-    private ExpressionSyntax ExtractReceiver(ExpressionSyntax expression)
-    {
-        if (expression is MemberAccessExpressionSyntax memberAccess)
-        {
-            return memberAccess.Expression;
-        }
-
-        return SyntaxFactory.IdentifierName(TransformWorkerProgramMarker.InstanceParameterName);
-    }
-
-    // Why not Visit synthetic nodes: GetSymbolInfo requires nodes from the original SemanticModel
-    // tree. Bare-member rewrite invents IdentifierName(InstanceParameterName), which must not be re-visited.
-    private ExpressionSyntax VisitReceiver(ExpressionSyntax receiver)
-    {
-        if (receiver.SyntaxTree != _semanticModel.SyntaxTree)
-        {
-            return receiver;
-        }
-
-        return (ExpressionSyntax)Visit(receiver);
-    }
-
-    private (bool owned, bool isStatic, INamedTypeSymbol containingType) ResolveOwnedMember(ISymbol symbol)
-    {
-        INamedTypeSymbol containingType;
-        bool isStatic;
-
-        if (symbol is IMethodSymbol methodSymbol)
-        {
-            if (methodSymbol.IsExtensionMethod)
-            {
-                return (false, false, null);
-            }
-
-            containingType = methodSymbol.ContainingType;
-            isStatic = methodSymbol.IsStatic;
-        }
-        else if (symbol is IFieldSymbol fieldSymbol)
-        {
-            containingType = fieldSymbol.ContainingType;
-            isStatic = fieldSymbol.IsStatic;
-        }
-        else if (symbol is IPropertySymbol propertySymbol)
-        {
-            containingType = propertySymbol.ContainingType;
-            isStatic = propertySymbol.IsStatic;
-        }
-        else if (symbol is IEventSymbol eventSymbol)
-        {
-            containingType = eventSymbol.ContainingType;
-            isStatic = eventSymbol.IsStatic;
-        }
-        else
-        {
-            return (false, false, null);
-        }
-
-        if (containingType == null)
-        {
-            return (false, false, null);
-        }
-
-        if (!IsInInheritanceHierarchy(_targetType, containingType))
-        {
-            return (false, false, null);
-        }
-
-        return (true, isStatic, containingType);
-    }
-
-    private static bool IsInInheritanceHierarchy(INamedTypeSymbol derived, INamedTypeSymbol candidate)
-    {
-        for (INamedTypeSymbol current = derived; current != null; current = current.BaseType)
-        {
-            if (SymbolEqualityComparer.Default.Equals(current, candidate))
-            {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private static bool IsMemberAccessNameSide(SimpleNameSyntax node)
-    {
-        return node.Parent is MemberAccessExpressionSyntax memberAccess
-            && memberAccess.Name == node;
-    }
-
-    private static bool IsQualifiedNameRightSide(SimpleNameSyntax node)
-    {
-        return node.Parent is QualifiedNameSyntax qualifiedName
-            && qualifiedName.Right == node;
-    }
-
-    private static bool IsMemberBindingName(SimpleNameSyntax node)
-    {
-        return node.Parent is MemberBindingExpressionSyntax memberBinding
-            && memberBinding.Name == node;
-    }
-
-    // `new T { _field = 1 }` must keep the bare member name; qualifying to instance._field is
-    // invalid inside an object/collection initializer.
-    private static bool IsObjectOrCollectionInitializerMemberName(SimpleNameSyntax node)
-    {
-        if (node.Parent is not AssignmentExpressionSyntax assignment || assignment.Left != node)
-        {
-            return false;
-        }
-
-        return assignment.Parent is InitializerExpressionSyntax;
-    }
-}
-
-// Nested access to the instance parameter name without exposing TransformWorkerProgram fields
-// to the rewriter as a circular partial. Kept as a tiny marker type so the rewriter stays free
-// of string literals scattered across Visit overrides.
-internal static class TransformWorkerProgramMarker
-{
-    // Why "__uloopInstance": the shim prepends this receiver parameter to the user's own
-    // parameter list verbatim, so a plain name like "instance" collides (CS0100) with any
-    // user parameter or local of that name. The uloop-prefixed name makes collisions
-    // practically impossible.
-    public const string InstanceParameterName = "__uloopInstance";
-
-    // Keep in sync with HotReloadAddedFieldStore in ToolContracts.
-    public const string AddedFieldStoreTypeName =
-        "global::io.github.hatayama.UnityCliLoop.ToolContracts.HotReloadAddedFieldStore";
-
-    public const string AddedFieldGetOrInitMethodName = "GetOrInit";
-
-    public const string AddedFieldSetMethodName = "Set";
-
-    public const string AddedFieldGetOrInitStaticMethodName = "GetOrInitStatic";
-
-    public const string AddedFieldSetStaticMethodName = "SetStatic";
-
-    // Keep in sync with HotReloadAddedFieldStore.FieldKeySeparator.
-    public const string AddedFieldKeySeparator = "::";
-}
-
-internal static class ShimMethodFactory
-{
-    public static MethodDeclarationSyntax ToShimMethod(
-        MethodDeclarationSyntax rewrittenOriginal,
-        IMethodSymbol methodSymbol)
-    {
-        TypeSyntax returnType = rewrittenOriginal.ReturnType.WithoutTrivia();
-        SyntaxTokenList modifiers = SyntaxFactory.TokenList(
-            SyntaxFactory.Token(SyntaxKind.PublicKeyword),
-            SyntaxFactory.Token(SyntaxKind.StaticKeyword));
-
-        // Async is preserved so the shim assembly still emits a state machine when the original
-        // was async (transplant covers the stub; MoveNext stays in the shim assembly).
-        if (rewrittenOriginal.Modifiers.Any(static modifier => modifier.IsKind(SyntaxKind.AsyncKeyword)))
-        {
-            modifiers = modifiers.Add(SyntaxFactory.Token(SyntaxKind.AsyncKeyword));
-        }
-
-        SeparatedSyntaxList<ParameterSyntax> parameters = BuildShimParameters(rewrittenOriginal, methodSymbol);
-        MethodDeclarationSyntax shim = rewrittenOriginal
-            .WithAttributeLists(default)
-            .WithModifiers(modifiers)
-            .WithReturnType(returnType)
-            .WithParameterList(SyntaxFactory.ParameterList(parameters))
-            .WithExplicitInterfaceSpecifier(null)
-            .WithConstraintClauses(default)
-            .WithLeadingTrivia(StripDirectiveTrivia(rewrittenOriginal.GetLeadingTrivia()))
-            .WithTrailingTrivia(StripDirectiveTrivia(rewrittenOriginal.GetTrailingTrivia()));
-
-        // Expression-bodied methods must keep their terminating semicolon; block bodies must not.
-        return rewrittenOriginal.ExpressionBody != null
-            ? shim.WithSemicolonToken(SyntaxFactory.Token(SyntaxKind.SemicolonToken))
-            : shim.WithSemicolonToken(default);
-    }
-
-    // Why strip directives: #if sits on the method's leading trivia while its matching #endif
-    // belongs to the next token, so copied directives are unbalanced in the shim; #line mapping
-    // is injected later from annotations and needs no user directives.
-    private static SyntaxTriviaList StripDirectiveTrivia(SyntaxTriviaList trivia)
-    {
-        List<SyntaxTrivia> kept = new List<SyntaxTrivia>();
-        foreach (SyntaxTrivia item in trivia)
-        {
-            if (!item.IsDirective)
-            {
-                kept.Add(item);
-            }
-        }
-
-        return SyntaxFactory.TriviaList(kept);
-    }
-
-    private static SeparatedSyntaxList<ParameterSyntax> BuildShimParameters(
-        MethodDeclarationSyntax rewrittenOriginal,
-        IMethodSymbol methodSymbol)
-    {
-        List<ParameterSyntax> parameters = new List<ParameterSyntax>();
-        if (!methodSymbol.IsStatic)
-        {
-            TypeSyntax instanceType = SyntaxFactory.ParseTypeName(
-                methodSymbol.ContainingType.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat));
-            parameters.Add(
-                SyntaxFactory.Parameter(SyntaxFactory.Identifier(TransformWorkerProgramMarker.InstanceParameterName))
-                    .WithType(instanceType));
-        }
-
-        foreach (ParameterSyntax originalParameter in rewrittenOriginal.ParameterList.Parameters)
-        {
-            parameters.Add(originalParameter.WithoutTrivia());
-        }
-
-        return SyntaxFactory.SeparatedList(parameters);
-    }
-}
-
-internal sealed class ShimTypeBuilder
-{
-    private readonly List<MethodDeclarationSyntax> _methods = new List<MethodDeclarationSyntax>();
-
-    public ShimTypeBuilder(
-        string shimTypeName,
-        string namespaceName,
-        List<UsingDirectiveSyntax> usings)
-    {
-        ShimTypeName = shimTypeName;
-        NamespaceName = namespaceName ?? string.Empty;
-        Usings = usings ?? new List<UsingDirectiveSyntax>();
-        AccessorPlan = new AccessorPlan();
-    }
-
-    public string ShimTypeName { get; }
-
-    public string NamespaceName { get; }
-
-    public List<UsingDirectiveSyntax> Usings { get; }
-
-    /// <summary>
-    /// Shim-type-level accessor registry — shared across all delegation methods in this type so
-    /// AllocateName stays unique and overloads cannot collide after a per-method merge.
-    /// </summary>
-    public AccessorPlan AccessorPlan { get; }
-
-    public void AddMethod(MethodDeclarationSyntax shimMethod, string shimMethodName)
-    {
-        MethodDeclarationSyntax named = shimMethod.WithIdentifier(SyntaxFactory.Identifier(shimMethodName));
-        _methods.Add(named);
-    }
-
-    public IEnumerable<MemberDeclarationSyntax> EmitMembers()
-    {
-        foreach (AccessorEntry accessor in AccessorPlan.Entries)
-        {
-            yield return accessor.EmitFieldDeclaration();
-        }
-
-        if (AccessorPlan.Entries.Count > 0)
-        {
-            yield return EmitBindAccessorsMethod();
-        }
-
-        foreach (MethodDeclarationSyntax method in _methods)
-        {
-            yield return method;
-        }
-    }
-
-    private MethodDeclarationSyntax EmitBindAccessorsMethod()
-    {
-        List<StatementSyntax> statements = new List<StatementSyntax>();
-        foreach (AccessorEntry accessor in AccessorPlan.Entries)
-        {
-            statements.Add(accessor.EmitBindStatement());
-        }
-
-        return SyntaxFactory.MethodDeclaration(
-                SyntaxFactory.PredefinedType(SyntaxFactory.Token(SyntaxKind.VoidKeyword)),
-                "__BindAccessors")
-            .WithModifiers(
-                SyntaxFactory.TokenList(
-                    SyntaxFactory.Token(SyntaxKind.PublicKeyword),
-                    SyntaxFactory.Token(SyntaxKind.StaticKeyword)))
-            .WithBody(SyntaxFactory.Block(statements));
-    }
-}
-
-internal static class ShimSourceEmitter
-{
-    public static string Emit(
-        CompilationUnitSyntax originalRoot,
-        List<ShimTypeBuilder> shimTypes,
-        string projectRelativePath)
-    {
-        if (shimTypes.Count == 0)
-        {
-            return string.Empty;
-        }
-
-        // projectRelativePath shape is validated at TransformFile's input boundary (ParseErrors).
-
-        // Emit each shim type in the original type's namespace (and with that type's usings) so
-        // unqualified sibling-type references in transplanted bodies still resolve. Manifest
-        // shimTypeName stays the short name; orchestrator resolves by Type.Name.
-        CompilationUnitSyntax unit = SyntaxFactory.CompilationUnit();
-        foreach (ShimTypeBuilder shimType in shimTypes)
-        {
-            ClassDeclarationSyntax classDeclaration = SyntaxFactory.ClassDeclaration(shimType.ShimTypeName)
-                .WithModifiers(
-                    SyntaxFactory.TokenList(
-                        SyntaxFactory.Token(SyntaxKind.PublicKeyword),
-                        SyntaxFactory.Token(SyntaxKind.StaticKeyword)))
-                .WithMembers(SyntaxFactory.List(shimType.EmitMembers()));
-
-            if (string.IsNullOrEmpty(shimType.NamespaceName))
-            {
-                foreach (UsingDirectiveSyntax usingDirective in shimType.Usings)
-                {
-                    unit = unit.AddUsings(usingDirective);
-                }
-
-                unit = unit.AddMembers(classDeclaration);
-            }
-            else
-            {
-                NamespaceDeclarationSyntax namespaceDeclaration = SyntaxFactory.NamespaceDeclaration(
-                        SyntaxFactory.ParseName(shimType.NamespaceName))
-                    .WithUsings(SyntaxFactory.List(shimType.Usings))
-                    .WithMembers(
-                        SyntaxFactory.SingletonList<MemberDeclarationSyntax>(classDeclaration));
-                unit = unit.AddMembers(namespaceDeclaration);
-            }
-        }
-
-        // Why after NormalizeWhitespace: formatting would otherwise shift #line relative to
-        // statements; annotations survive formatting so we inject directives on the final tree.
-        unit = unit.NormalizeWhitespace();
-        unit = InjectLineDirectives(unit, projectRelativePath);
-
-        StringBuilder builder = new StringBuilder();
-        builder.AppendLine(
-            "// Generated shims mirror user method signatures verbatim; repo style rules apply to hand-written code only.");
-        builder.Append(unit.ToFullString());
-        return builder.ToString();
-    }
-
-    private static CompilationUnitSyntax InjectLineDirectives(
-        CompilationUnitSyntax unit,
-        string projectRelativePath)
-    {
-        List<SyntaxNode> annotatedNodes = unit.GetAnnotatedNodes(TransformWorkerProgram.UloopLineAnnotationKind)
-            .ToList();
-        if (annotatedNodes.Count > 0)
-        {
-            unit = unit.ReplaceNodes(
-                annotatedNodes,
-                (original, rewritten) =>
-                {
-                    SyntaxAnnotation annotation = original
-                        .GetAnnotations(TransformWorkerProgram.UloopLineAnnotationKind)
-                        .First();
-                    // Why leading trivia starts/ends with newline: #line must occupy its own line.
-                    string directiveText =
-                        "\n#line " + annotation.Data + " \"" + projectRelativePath + "\"\n";
-                    SyntaxTriviaList leading = SyntaxFactory.ParseLeadingTrivia(directiveText);
-                    return rewritten.WithLeadingTrivia(leading.AddRange(rewritten.GetLeadingTrivia()));
-                });
-        }
-
-        // Reset mapping after each method so scaffold (__BindAccessors, fields, class braces)
-        // does not inherit the previous method's document/line.
-        List<MethodDeclarationSyntax> methods = unit.DescendantNodes()
-            .OfType<MethodDeclarationSyntax>()
-            .ToList();
-        if (methods.Count == 0)
-        {
-            return unit;
-        }
-
-        // Why ParseLeadingTrivia into trailing: ParseTrailingTrivia does not reliably produce
-        // LineDirectiveTrivia for "#line default", while ParseLeadingTrivia does — and directive
-        // trivia is legal in a trailing trivia list for ToFullString emission.
-        return unit.ReplaceNodes(
-            methods,
-            (original, rewritten) =>
-            {
-                SyntaxTriviaList defaultDirective = SyntaxFactory.ParseLeadingTrivia("\n#line default\n");
-                return rewritten.WithTrailingTrivia(
-                    rewritten.GetTrailingTrivia().AddRange(defaultDirective));
-            });
-    }
-}
-
-internal static class CecilTypeNames
-{
-    public static string ToMetadataName(INamedTypeSymbol typeSymbol)
-    {
-        if (typeSymbol.ContainingType != null)
-        {
-            return ToMetadataName(typeSymbol.ContainingType) + "/" + typeSymbol.MetadataName;
-        }
-
-        if (typeSymbol.ContainingNamespace == null || typeSymbol.ContainingNamespace.IsGlobalNamespace)
-        {
-            return typeSymbol.MetadataName;
-        }
-
-        return typeSymbol.ContainingNamespace.ToDisplayString() + "." + typeSymbol.MetadataName;
-    }
-
-    public static string ToParameterTypeFullName(IParameterSymbol parameterSymbol)
-    {
-        // Roslyn's IParameterSymbol.Type omits the byref wrapper; Cecil FullName uses a trailing '&'.
-        string typeName = ToCecilFullName(parameterSymbol.Type);
-        if (parameterSymbol.RefKind != RefKind.None)
-        {
-            return typeName + "&";
-        }
-
-        return typeName;
-    }
-
-    public static string ToCecilFullName(ITypeSymbol typeSymbol)
-    {
-        // Why here (not only the parameter top level): source `List<dynamic>` / `dynamic[]`
-        // nest TypeKind.Dynamic while compiled metadata uses System.Object at the same depth.
-        if (typeSymbol != null && typeSymbol.TypeKind == TypeKind.Dynamic)
-        {
-            return "System.Object";
-        }
-
-        if (typeSymbol is IPointerTypeSymbol pointerType)
-        {
-            return ToCecilFullName(pointerType.PointedAtType) + "*";
-        }
-
-        if (typeSymbol is IArrayTypeSymbol arrayType)
-        {
-            string elementName = ToCecilFullName(arrayType.ElementType);
-            if (arrayType.Rank == 1)
-            {
-                return elementName + "[]";
-            }
-
-            // Cecil ArrayType.FullName for non-vector arrays uses "[0...,0...]" (lower bounds),
-            // not the C# syntactic "[,]".
-            string[] dimensionMarks = new string[arrayType.Rank];
-            for (int index = 0; index < arrayType.Rank; index++)
-            {
-                dimensionMarks[index] = "0...";
-            }
-
-            return elementName + "[" + string.Join(",", dimensionMarks) + "]";
-        }
-
-        if (typeSymbol is INamedTypeSymbol namedType)
-        {
-            // Cecil nests constructed generics as Outer/Inner<all-args-outer-to-inner>, so collect
-            // TypeArguments from the containment chain rather than only the leaf type.
-            List<ITypeSymbol> typeArguments = new List<ITypeSymbol>();
-            CollectConstructedTypeArgumentsOuterToInner(namedType, typeArguments);
-            string head = ToMetadataName(namedType.OriginalDefinition);
-            if (typeArguments.Count == 0)
-            {
-                return head;
-            }
-
-            string args = string.Join(",", typeArguments.Select(ToCecilFullName));
-            return head + "<" + args + ">";
-        }
-
-        return typeSymbol.ToDisplayString(
-            new SymbolDisplayFormat(
-                typeQualificationStyle: SymbolDisplayTypeQualificationStyle.NameAndContainingTypesAndNamespaces));
-    }
-
-    private static void CollectConstructedTypeArgumentsOuterToInner(
-        INamedTypeSymbol namedType,
-        List<ITypeSymbol> typeArguments)
-    {
-        if (namedType.ContainingType != null)
-        {
-            CollectConstructedTypeArgumentsOuterToInner(namedType.ContainingType, typeArguments);
-        }
-
-        if (namedType.IsGenericType && !namedType.IsUnboundGenericType)
-        {
-            typeArguments.AddRange(namedType.TypeArguments);
-        }
-    }
-}
-
-// What: syntax-key maps for one file's baseline-vs-current comparison during TransformFile.
-internal sealed class BaselineSnapshotState
-{
-    public bool HasBaseline { get; set; }
-
-    public bool BaselineDisabledByDuplicateKeys { get; set; }
-
-    public CompilationUnitSyntax SnapshotRoot { get; set; }
-
-    public Dictionary<string, MethodDeclarationSyntax> SnapshotMethodMap { get; set; }
-
-    public Dictionary<string, MethodDeclarationSyntax> PlainCurrentMethodMap { get; set; }
-
-    public Dictionary<string, PropertyDeclarationSyntax> SnapshotPropertyMap { get; set; }
-
-    public Dictionary<string, IndexerDeclarationSyntax> SnapshotIndexerMap { get; set; }
-
-    public Dictionary<string, ConstructorDeclarationSyntax> SnapshotConstructorMap { get; set; }
-
-    public Dictionary<string, MemberDeclarationSyntax> SnapshotOperatorMap { get; set; }
-
-    public Dictionary<string, EventDeclarationSyntax> SnapshotEventMap { get; set; }
-
-    public Dictionary<string, PropertyDeclarationSyntax> PlainCurrentPropertyMap { get; set; }
-
-    public Dictionary<string, IndexerDeclarationSyntax> PlainCurrentIndexerMap { get; set; }
-
-    public Dictionary<string, ConstructorDeclarationSyntax> PlainCurrentConstructorMap { get; set; }
-
-    public Dictionary<string, MemberDeclarationSyntax> PlainCurrentOperatorMap { get; set; }
-
-    public Dictionary<string, EventDeclarationSyntax> PlainCurrentEventMap { get; set; }
-
-    public (
-        Dictionary<string, PropertyDeclarationSyntax> SnapshotPropertyMap,
-        Dictionary<string, IndexerDeclarationSyntax> SnapshotIndexerMap,
-        Dictionary<string, PropertyDeclarationSyntax> PlainCurrentPropertyMap,
-        Dictionary<string, IndexerDeclarationSyntax> PlainCurrentIndexerMap)
-        GetAccessorBaselineMaps()
-    {
-        if (!HasBaseline)
-        {
-            return (null, null, null, null);
-        }
-
-        return (SnapshotPropertyMap, SnapshotIndexerMap, PlainCurrentPropertyMap, PlainCurrentIndexerMap);
-    }
-
-    public (
-        Dictionary<string, ConstructorDeclarationSyntax> SnapshotConstructorMap,
-        Dictionary<string, MemberDeclarationSyntax> SnapshotOperatorMap,
-        Dictionary<string, EventDeclarationSyntax> SnapshotEventMap,
-        Dictionary<string, ConstructorDeclarationSyntax> PlainCurrentConstructorMap,
-        Dictionary<string, MemberDeclarationSyntax> PlainCurrentOperatorMap,
-        Dictionary<string, EventDeclarationSyntax> PlainCurrentEventMap)
-        GetUnsupportedMemberBaselineMaps()
-    {
-        if (!HasBaseline)
-        {
-            return (null, null, null, null, null, null);
-        }
-
-        return (
-            SnapshotConstructorMap,
-            SnapshotOperatorMap,
-            SnapshotEventMap,
-            PlainCurrentConstructorMap,
-            PlainCurrentOperatorMap,
-            PlainCurrentEventMap);
-    }
-}
-
-internal sealed class TypeEmitState
-{
-    public TypeDeclarationSyntax TypeDeclaration { get; set; }
-
-    public INamedTypeSymbol TypeSymbol { get; set; }
-
-    public string TypeMetadataNameFromSyntax { get; set; }
-
-    public ShimTypeBuilder CurrentShimType { get; set; }
-
-    public List<QueuedShimMethod> QueuedMethods { get; } = new List<QueuedShimMethod>();
-
-    public bool TypeIsAbsentFromCompiledAssembly { get; set; }
-}
-
-internal sealed class QueuedShimMethod
-{
-    public MethodDeclarationSyntax MethodDeclaration { get; set; }
-
-    public IMethodSymbol MethodSymbol { get; set; }
-
-    public MethodTransformDecision Decision { get; set; }
-
-    public string ShimMethodName { get; set; }
-
-    public ShimTypeBuilder ShimType { get; set; }
-
-    public int SourceStartLine { get; set; }
-
-    public int SourceEndLine { get; set; }
-
-    public string[] ParameterTypeFullNames { get; set; }
-
-    public string MethodKey { get; set; }
-
-    public bool IsAddedMethod { get; set; }
-
-    public bool ReplacesCompiledMethod { get; set; }
-}
-
-internal sealed class AddedMethodBinding
-{
-    public string MethodKey { get; set; }
-
-    public string ShimTypeName { get; set; }
-
-    public string ShimMethodName { get; set; }
-
-    public string NamespaceName { get; set; }
-
-    public bool IsStatic { get; set; }
-}
-
-internal sealed class AddedMethodCatalog
-{
-    private readonly Dictionary<string, AddedMethodBinding> _byKey =
-        new Dictionary<string, AddedMethodBinding>(StringComparer.Ordinal);
-    private readonly HashSet<string> _classifiedAddedKeys = new HashSet<string>(StringComparer.Ordinal);
-    private readonly HashSet<string> _addedSyntaxKeys = new HashSet<string>(StringComparer.Ordinal);
-    private readonly HashSet<string> _addedTypeSyntaxKeys = new HashSet<string>(StringComparer.Ordinal);
-    private readonly HashSet<string> _addedPropertySyntaxKeys = new HashSet<string>(StringComparer.Ordinal);
-    private readonly HashSet<string> _removedSyntaxKeys = new HashSet<string>(StringComparer.Ordinal);
-
-    public IReadOnlyCollection<string> AddedSyntaxKeys => _addedSyntaxKeys;
-
-    public IReadOnlyCollection<string> AddedTypeSyntaxKeys => _addedTypeSyntaxKeys;
-
-    public IReadOnlyCollection<string> AddedPropertySyntaxKeys => _addedPropertySyntaxKeys;
-
-    public IReadOnlyCollection<string> RemovedSyntaxKeys => _removedSyntaxKeys;
-
-    public void Register(AddedMethodBinding binding)
-    {
-        _byKey[binding.MethodKey] = binding;
-        MarkClassifiedAdded(binding.MethodKey);
-    }
-
-    public void MarkClassifiedAdded(string methodKey)
-    {
-        if (methodKey != null)
-        {
-            _classifiedAddedKeys.Add(methodKey);
-        }
-    }
-
-    public bool IsClassifiedAdded(string methodKey)
-    {
-        return methodKey != null && _classifiedAddedKeys.Contains(methodKey);
-    }
-
-    public bool IsUnavailableAdded(string methodKey)
-    {
-        return IsClassifiedAdded(methodKey) && !Contains(methodKey);
-    }
-
-    public void AddAddedSyntaxKey(string syntaxKey)
-    {
-        _addedSyntaxKeys.Add(syntaxKey);
-    }
-
-    public void AddAddedTypeSyntaxKey(string typeSyntaxKey)
-    {
-        if (typeSyntaxKey != null)
-        {
-            _addedTypeSyntaxKeys.Add(typeSyntaxKey);
-        }
-    }
-
-    public void AddAddedPropertySyntaxKey(string propertySyntaxKey)
-    {
-        if (propertySyntaxKey != null)
-        {
-            _addedPropertySyntaxKeys.Add(propertySyntaxKey);
-        }
-    }
-
-    public void AddRemovedSyntaxKey(string syntaxKey)
-    {
-        _removedSyntaxKeys.Add(syntaxKey);
-    }
-
-    public bool Contains(string methodKey)
-    {
-        return methodKey != null && _byKey.ContainsKey(methodKey);
-    }
-
-    public AddedMethodBinding FindOrNull(string methodKey)
-    {
-        if (methodKey == null)
-        {
-            return null;
-        }
-
-        return _byKey.TryGetValue(methodKey, out AddedMethodBinding binding) ? binding : null;
-    }
-
-    public void Unregister(string methodKey)
-    {
-        if (methodKey != null)
-        {
-            _byKey.Remove(methodKey);
-        }
-    }
-}
-
-/// <summary>
-/// What: one added field's store/const/unavailable binding used by skip evaluation and rewrite.
-/// </summary>
-internal sealed class AddedFieldBinding
-{
-    public string FieldKey { get; set; }
-
-    public string SyntaxKey { get; set; }
-
-    public string FieldName { get; set; }
-
-    public ITypeSymbol FieldType { get; set; }
-
-    public bool IsStatic { get; set; }
-
-    public bool IsConst { get; set; }
-
-    public object ConstantValue { get; set; }
-
-    public ExpressionSyntax Initializer { get; set; }
-
-    public string UnavailableReason { get; set; }
-
-    public bool IsStoreRewriteable => UnavailableReason == null && !IsConst;
-}
-
-/// <summary>
-/// What: file-wide catalog of added fields, syntax keys for drift strip, store-rewrite
-/// presence, and display names of fields rewritten in emitted shim bodies.
-/// </summary>
-internal sealed class AddedFieldCatalog
-{
-    private readonly Dictionary<string, AddedFieldBinding> _byKey =
-        new Dictionary<string, AddedFieldBinding>(StringComparer.Ordinal);
-    private readonly HashSet<string> _classifiedAddedKeys = new HashSet<string>(StringComparer.Ordinal);
-    private readonly HashSet<string> _rewrittenAddedFieldKeys = new HashSet<string>(StringComparer.Ordinal);
-    private readonly HashSet<string> _addedSyntaxKeys = new HashSet<string>(StringComparer.Ordinal);
-    private readonly HashSet<string> _removedSyntaxKeys = new HashSet<string>(StringComparer.Ordinal);
-
-    public IReadOnlyCollection<string> AddedSyntaxKeys => _addedSyntaxKeys;
-
-    public IReadOnlyCollection<string> RemovedSyntaxKeys => _removedSyntaxKeys;
-
-    public bool HasClassifiedAdded => _classifiedAddedKeys.Count > 0;
-
-    public bool HasStoreRewrites { get; private set; }
-
-    public void MarkClassifiedAdded(string fieldKey)
-    {
-        if (fieldKey != null)
-        {
-            _classifiedAddedKeys.Add(fieldKey);
-        }
-    }
-
-    public void AddAddedSyntaxKey(string syntaxKey)
-    {
-        _addedSyntaxKeys.Add(syntaxKey);
-    }
-
-    public void AddRemovedSyntaxKey(string syntaxKey)
-    {
-        _removedSyntaxKeys.Add(syntaxKey);
-    }
-
-    public void RegisterStore(AddedFieldBinding binding)
-    {
-        _byKey[binding.FieldKey] = binding;
-        MarkClassifiedAdded(binding.FieldKey);
-    }
-
-    public void RegisterConst(AddedFieldBinding binding)
-    {
-        _byKey[binding.FieldKey] = binding;
-        MarkClassifiedAdded(binding.FieldKey);
-    }
-
-    // Why rewritten keys, not RegisterStore/RegisterConst: those fire at declaration
-    // classification, so unused fields and isolation-excluded bodies would still list.
-    // Excluded methods are dropped in QueueTypeMethods before rewrite, so a file-wide
-    // rewrite set matches emitted entries without per-entry tracking.
-    public string[] ListRewrittenAddedFieldDisplayNames()
-    {
-        List<string> names = new List<string>(_rewrittenAddedFieldKeys.Count);
-        foreach (string fieldKey in _rewrittenAddedFieldKeys)
-        {
-            names.Add(FormatAddedFieldDisplayName(fieldKey));
-        }
-
-        names.Sort(StringComparer.Ordinal);
-        return names.ToArray();
-    }
-
-    // Why this shape: method labels replace '/' with '+' then join with '.', so field
-    // names stay comparable to Methods[].Method (Ns.Type.field).
-    private static string FormatAddedFieldDisplayName(string fieldKey)
-    {
-        int separatorIndex = fieldKey.IndexOf(
-            TransformWorkerProgramMarker.AddedFieldKeySeparator,
-            StringComparison.Ordinal);
-        Debug.Assert(
-            separatorIndex >= 0,
-            "fieldKey is always built with FormatAddedFieldStoreKey / BuildSyntaxFieldKey.");
-
-        string typeMetadataName = fieldKey.Substring(0, separatorIndex).Replace('/', '+');
-        string fieldName = fieldKey.Substring(
-            separatorIndex + TransformWorkerProgramMarker.AddedFieldKeySeparator.Length);
-        return typeMetadataName + "." + fieldName;
-    }
-
-    public void RegisterUnavailable(AddedFieldBinding binding)
-    {
-        _byKey[binding.FieldKey] = binding;
-        MarkClassifiedAdded(binding.FieldKey);
-    }
-
-    public AddedFieldBinding FindOrNull(string fieldKey)
-    {
-        if (fieldKey == null)
-        {
-            return null;
-        }
-
-        return _byKey.TryGetValue(fieldKey, out AddedFieldBinding binding) ? binding : null;
-    }
-
-    public void MarkStoreRewrite(string fieldKey)
-    {
-        Debug.Assert(!string.IsNullOrEmpty(fieldKey), "fieldKey must not be null or empty.");
-        HasStoreRewrites = true;
-        _rewrittenAddedFieldKeys.Add(fieldKey);
-    }
-
-    public void MarkConstFold(string fieldKey)
-    {
-        Debug.Assert(!string.IsNullOrEmpty(fieldKey), "fieldKey must not be null or empty.");
-        _rewrittenAddedFieldKeys.Add(fieldKey);
-    }
-}
-
-/// <summary>
-/// What: skip and warning strings for added-field classification. Keep in the existing
-/// "reason + run uloop compile" style; the worker cannot reference HotReloadConstants.
-/// </summary>
-internal static class AddedFieldSkipReasons
-{
-    public const string StructHost =
-        "Added fields on struct types are skipped; the store requires a reference-type instance. "
-        + "Run 'uloop compile' to add them.";
-
-    public const string InitializerNotLiteralOrExternalStatic =
-        "Added field initializer is not a literal or externally visible static member; "
-        + "the shim lambda cannot use instance, host-type, or same-file added members. "
-        + "Run 'uloop compile'.";
-
-    public const string FieldTypeNotExternallyVisible =
-        "Added field type is not visible to the shim assembly. Run 'uloop compile'.";
-
-    public const string IncrementNotNumeric =
-        "Increment or decrement of an added field is skipped unless the type is a numeric "
-        + "primitive or enum. Run 'uloop compile'.";
-
-    public const string RefOutIn =
-        "Added fields cannot be passed by ref, out, or in. Run 'uloop compile'.";
-
-    public const string ConsumedWrite =
-        "The value of an assignment to an added field is consumed; the store write returns void. "
-        + "Run 'uloop compile'.";
-
-    public const string DoubleEvalReceiver =
-        "Assignment to an added field would evaluate a receiver with possible side effects twice. "
-        + "Run 'uloop compile'.";
-
-    public const string ValueTypeMemberWrite =
-        "Writes to members of an added value-type field, and instance method calls on that field, "
-        + "cannot be rewritten. Run 'uloop compile'.";
-
-    public const string UnavailableAddedField =
-        "Uses an added field that hot reload cannot emit. Run 'uloop compile'.";
-
-    public const string FieldTypeChanged =
-        "Field '{0}' has a different type in the compiled assembly. Run 'uloop compile'.";
-
-    public const string FieldModifiersChanged =
-        "Field '{0}' changed its static or const modifier in the compiled assembly. Run 'uloop compile'.";
-
-    public const string MemberKindChanged =
-        "Field '{0}' is declared as a property or an event in the compiled assembly. Run 'uloop compile'.";
-
-    public const string SerializeWarningFormat =
-        "Added field '{0}' has a serialization attribute, so it will not appear in the Inspector "
-        + "or serialize until 'uloop compile'.";
-}
-
-internal static class AddedMethodSkipReasons
-{
-    public const string VirtualOrAbstract =
-        "Added virtual, override, or abstract methods are skipped; the compiled type has no vtable slot. "
-        + "Run 'uloop compile' to add them.";
-
-    public const string AddedProperty =
-        "Added properties are out of scope for hot reload; the compiled assembly has no such member. "
-        + "Use a 'const' or a plain added field for the value, or run 'uloop compile'.";
-
-    public const string Generic =
-        "Added generic methods are skipped; hot reload cannot emit a typed shim for them. "
-        + "Run 'uloop compile'.";
-
-    public const string MethodGroupReference =
-        "Methods that capture an added method as a method group or delegate are skipped; "
-        + "the shim signature does not match. Run 'uloop compile'.";
-
-    public const string ConditionalAccess =
-        "Added-method calls through conditional access are skipped; there is no rewrite shape. "
-        + "Run 'uloop compile'.";
-
-    public const string UnavailableAddedCall =
-        "Calls an added method that hot reload cannot emit. Run 'uloop compile'.";
-
-    public const string NewTypeOutOfScope =
-        "New types are out of scope for hot reload; run 'uloop compile' to add them.";
-
-    public const string InterfaceMember =
-        "Interface members are not patchable. Run 'uloop compile'.";
-
-    public const string InaccessibleAccessNoRewrite =
-        "Added methods whose bodies access private/internal members are skipped when the access "
-        + "has no accessor rewrite (the added method JIT-compiles normally and fails accessibility "
-        + "checks).";
-}
-
-internal static class UnityMessageNames
-{
-    // Keep in sync with the Unity-message set that PR-5 will document in
-    // Packages/src/Editor/FirstPartyTools/HotReload/Skill/SKILL.md.
-    public const string AddedMessageWarningFormat =
-        "Added Unity message '{0}' on {1} will not be invoked by the engine until 'uloop compile'; "
-        + "Unity discovers messages by reflection on the compiled type.";
-
-    private static readonly HashSet<string> Names = new HashSet<string>(StringComparer.Ordinal)
-    {
-        "Awake",
-        "Start",
-        "OnEnable",
-        "OnDisable",
-        "OnDestroy",
-        "Update",
-        "LateUpdate",
-        "FixedUpdate",
-        "OnGUI",
-        "Reset",
-        "OnValidate",
-        "OnCollisionEnter",
-        "OnCollisionExit",
-        "OnCollisionStay",
-        "OnTriggerEnter",
-        "OnTriggerExit",
-        "OnTriggerStay",
-        "OnCollisionEnter2D",
-        "OnCollisionExit2D",
-        "OnCollisionStay2D",
-        "OnTriggerEnter2D",
-        "OnTriggerExit2D",
-        "OnTriggerStay2D",
-        "OnMouseDown",
-        "OnMouseUp",
-        "OnMouseEnter",
-        "OnMouseExit",
-        "OnMouseOver",
-        "OnMouseDrag",
-        "OnBecameVisible",
-        "OnBecameInvisible",
-        "OnApplicationQuit",
-        "OnApplicationPause",
-        "OnApplicationFocus",
-        "OnTransformChildrenChanged",
-        "OnTransformParentChanged",
-        "OnRectTransformDimensionsChange",
-        "OnParticleCollision",
-        "OnParticleTrigger",
-        "OnControllerColliderHit",
-        "OnJointBreak",
-        "OnJointBreak2D",
-        "OnAnimatorMove",
-        "OnAnimatorIK",
-        "OnDrawGizmos",
-        "OnDrawGizmosSelected"
-    };
-
-    public static bool Contains(string methodName)
-    {
-        return methodName != null && Names.Contains(methodName);
-    }
-}
-
-internal sealed class WorkerInput
-{
-    public string SourcePath { get; set; }
-
-    public string[] Defines { get; set; }
-
-    public string[] ReferencePaths { get; set; }
-
-    public string TargetTypesAssemblyPath { get; set; }
-
-    // Method keys (see TransformWorkerProgram.BuildMethodKey) that the orchestrator already
-    // reported Failed from a first compile round; the retry excludes them so it does not fail
-    // on the same error again.
-    public string[] ExcludedMethodKeys { get; set; }
-
-    // Added-method keys whose shim bodies failed the first compile. Distinct from
-    // ExcludedMethodKeys so a healthy added shim is not dropped when an existing method fails.
-    public string[] ExcludedAddedMethodKeys { get; set; }
-
-    // Verified snapshot text for edited-method detection. Null = no baseline, patch all methods.
-    // Why pass text (not a path): avoids an IO race between orchestrator verification and worker
-    // read that would crash the whole file under the no-try-catch policy.
-    public string SnapshotSource { get; set; }
-
-    // Project-relative forward-slash path embedded in #line document names.
-    public string ProjectRelativePath { get; set; }
-
-    // Absolute paths of every source file in the edited file's compilation assembly.
-    // Null/omitted is treated as empty (no sibling global usings collected).
-    public string[] AssemblySourcePaths { get; set; }
-}
-
-internal sealed class WorkerOutput
-{
-    public string ShimSource { get; set; }
-
-    public WorkerEntry[] Entries { get; set; }
-
-    public WorkerSkipped[] Skipped { get; set; }
-
-    public string[] DeclarationDriftWarnings { get; set; }
-
-    public string[] ParseErrors { get; set; }
-
-    public WorkerUnchangedMethod[] UnchangedMethods { get; set; }
-
-    public bool BaselineDisabledByDuplicateKeys { get; set; }
-
-    public WorkerRemovedMember[] RemovedMembers { get; set; }
-
-    public WorkerRemovedMethodSignature[] RemovedMethodSignatures { get; set; }
-
-    public bool HasAccessorDelegates { get; set; }
-
-    // True when shim bodies rewrite added-field accesses to HotReloadAddedFieldStore.
-    // Keep in sync with TransformWorkerOutputDto.hasAddedFieldRewrites.
-    public bool HasAddedFieldRewrites { get; set; }
-
-    // Keep in sync with TransformWorkerOutputDto.addedFieldNames.
-    public string[] AddedFieldNames { get; set; }
-
-    // Keep in sync with TransformWorkerOutputDto.sourceContentSha256.
-    public string SourceContentSha256 { get; set; }
-}
-
-internal sealed class WorkerRemovedMember
-{
-    public string Kind { get; set; }
-
-    public string Name { get; set; }
-}
-
-internal sealed class WorkerRemovedMethodSignature
-{
-    public string TypeMetadataName { get; set; }
-
-    public string MethodName { get; set; }
-
-    public string[] ParameterTypeFullNames { get; set; }
-
-    public int GenericArity { get; set; }
-}
-
-internal enum CompiledMethodMatch
-{
-    NotFound,
-    Matched,
-    ReturnTypeChanged
-}
-
-internal enum CompiledFieldMatch
-{
-    NotFound,
-    Matched,
-    FieldTypeChanged,
-    FieldModifiersChanged,
-    MemberKindChanged
-}
-
-internal sealed class WorkerUnchangedMethod
-{
-    public string TypeMetadataName { get; set; }
-
-    public string MethodName { get; set; }
-
-    public string[] ParameterTypeFullNames { get; set; }
-
-    public int GenericArity { get; set; }
-}
-
-internal sealed class WorkerEntry
-{
-    public string TypeMetadataName { get; set; }
-
-    public string MethodName { get; set; }
-
-    public string[] ParameterTypeFullNames { get; set; }
-
-    public int GenericArity { get; set; }
-
-    public string ShimTypeName { get; set; }
-
-    public string ShimMethodName { get; set; }
-
-    // "transplant" | "delegation" | "addedMethod" — see PatchKinds.
-    public string PatchKind { get; set; }
-
-    // Method keys of added methods this entry's body invokes. Empty when none.
-    public string[] CalledAddedMethodKeys { get; set; }
-
-    // 1-based, both ends inclusive, within the original edited source file.
-    public int SourceStartLine { get; set; }
-
-    public int SourceEndLine { get; set; }
-
-    // Null when the method is not a one-shot lifecycle method and is not only called from them.
-    public string LifecycleNote { get; set; }
-
-    public bool ReplacesCompiledMethod { get; set; }
-}
-
-internal static class LifecycleNotes
-{
-    public static readonly string[] OneShotLifecycleMethodNames =
-    {
-        "Awake",
-        "Start",
-        "OnEnable",
-        "OnDisable",
-        "OnDestroy"
-    };
-
-    public const string DirectFormat =
-        "{0} is a one-shot lifecycle method; objects that already ran it will not run the "
-        + "patched body. It takes effect only for newly created objects.";
-}
-
-internal static class PatchKinds
-{
-    public const string Transplant = "transplant";
-    public const string Delegation = "delegation";
-
-    // Keep in sync with HotReloadConstants.PatchKindAddedMethod.
-    public const string AddedMethod = "addedMethod";
-}
-
-internal static class RemovedMemberKinds
-{
-    // Keep in sync with HotReloadConstants.RemovedMemberKindMethod / RemovedMemberKindField.
-    public const string Method = "method";
-
-    public const string Field = "field";
-}
-
-internal sealed class MethodTransformDecision
-{
-    public string SkipReason { get; private set; }
-
-    public string PatchKind { get; private set; }
-
-    public bool UsesDelegation { get; private set; }
-
-    public static MethodTransformDecision Skip(string reason)
-    {
-        return new MethodTransformDecision { SkipReason = reason };
-    }
-
-    public static MethodTransformDecision Transplant()
-    {
-        return new MethodTransformDecision { PatchKind = PatchKinds.Transplant };
-    }
-
-    public static MethodTransformDecision Delegation()
-    {
-        return new MethodTransformDecision
-        {
-            PatchKind = PatchKinds.Delegation,
-            UsesDelegation = true
-        };
-    }
-
-    public static MethodTransformDecision AddedMethod(bool usesDelegation)
-    {
-        return new MethodTransformDecision
-        {
-            PatchKind = PatchKinds.AddedMethod,
-            UsesDelegation = usesDelegation
-        };
-    }
-}
-
-internal sealed class WorkerSkipped
-{
-    public string Method { get; set; }
-
-    public string Reason { get; set; }
-
-    public string MethodKey { get; set; }
-
-    public string CalledAddedMethodKey { get; set; }
-}
